% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{accounting_point_eic}
\alias{accounting_point_eic}
\title{Get Accounting Point Z Energy Identification Codes}
\usage{
accounting_point_eic()
}
\value{
A tibble of accordingly filtered EIC codes, which contains such columns as
\code{eic_code}, \code{eic_display_name}, \code{eic_long_name}, \code{eic_parent},
\code{eic_responsible_party}, \code{eic_status}, \code{market_participant_postal_code},
\code{market_participant_iso_country_code}, \code{market_participant_vat_code},
\code{eic_type_function_list} and \code{type}.
}
\description{
This function downloads approved accounting point Z
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
An entity under balance responsibility where balance supplier change
can take place and for which commercial business processes are defined.
}
\examples{
\dontshow{if (there_is_provider()) withAutoprint(\{ # examplesIf}
eic_accounting_point <- entsoeapi::accounting_point_eic()

dplyr::glimpse(eic_accounting_point)
\dontshow{\}) # examplesIf}
}
