% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{all_allocated_eic}
\alias{all_allocated_eic}
\title{Get all Allocated Energy Identification Codes}
\usage{
all_allocated_eic()
}
\value{
A tibble of all allocated EIC codes, which contains such columns as
\code{revision_number}, \code{created_date_time}, \code{eic_code}, \code{doc_status_value},
\code{doc_status}, \code{instance_component_attribute}, \code{long_name}, \code{display_name},
\code{last_request_date}, \code{deactivation_requested_date_and_or_time_date},
\code{eic_code_market_participant_street_address},
\code{market_participant_vat_code_name}, \code{market_participant_acer_code_name},
\code{description}, \code{responsible_market_participant_mrid}, \code{function_names}
and \code{parent_market_document_mrid}
}
\description{
Beware, this is a REAL SLOW function, it runs for minutes, be patient!
This function downloads all allocated
energy identification codes from this link:
https://eepublicdownloads.blob.core.windows.net/
cio-lio/xml/allocated-eic-codes.xml
Further details are under:
https://www.entsoe.eu/data/energy-identification-codes-eic/
It is an alternative of \code{all_approved_eic()} function call
providing more details.
}
\examples{
\dontshow{if (there_is_provider()) withAutoprint(\{ # examplesIf}
eic_all <- entsoeapi::all_allocated_eic()

dplyr::glimpse(eic_all)
\dontshow{\}) # examplesIf}
}
