% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{all_eic}
\alias{all_eic}
\title{Get all EIC codes}
\usage{
all_eic()
}
\description{
This function downloads approved all Energy Identification Codes from
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes/
}
\examples{
eic_all <- all_eic()
}
