% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages_generation_units.R
\name{api_req_zip}
\alias{api_req_zip}
\title{save and unzip the zip file with xml data.}
\usage{
api_req_zip(url, file_type)
}
\arguments{
\item{url}{url.}

\item{file_type}{transmission or generation.}
}
\description{
save and unzip the zip file with xml data.
}
