% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{area_eic}
\alias{area_eic}
\title{Get Area Y Energy Identification Codes}
\usage{
area_eic()
}
\value{
A tibble of accordingly filtered EIC codes, which contains such columns as
\code{eic_code}, \code{eic_display_name}, \code{eic_long_name}, \code{eic_parent},
\code{eic_responsible_party}, \code{eic_status}, \code{market_participant_postal_code},
\code{market_participant_iso_country_code}, \code{market_participant_vat_code},
\code{eic_type_function_list} and \code{type}.
}
\description{
This function downloads approved area Y
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
}
\examples{
\dontshow{if (there_is_provider()) withAutoprint(\{ # examplesIf}
eic_area <- entsoeapi::area_eic()

dplyr::glimpse(eic_area)
\dontshow{\}) # examplesIf}
}
