% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{balancing_activated_reserves}
\alias{balancing_activated_reserves}
\title{Get Activated Balancing Reserves (17.1.E)}
\usage{
balancing_activated_reserves(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  reserve_type = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{reserve_type}{Defaults to NULL, otherwise choose among the
list of reserve type codes (A95, A96, A97, A98)
from business_types table}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
The amount of activated balancing energy.
}
\examples{

library(entsoeapi)
library(lubridate)

df <- balancing_activated_reserves(eic          = "10YHU-MAVIR----U",
                                   period_start = ymd(x = "2020-02-01",
                                                      tz = "CET"),
                                   period_end   = ymd(x = "2020-03-01",
                                                      tz = "CET"),
                                   tidy_output  = TRUE,
                                   reserve_type = "A96")
str(df)

}
