% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{congestion_income}
\alias{congestion_income}
\title{Get Implicit and Flow-based Allocations — Congestion Income (12.1.E)}
\usage{
congestion_income(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  contract_type = "A01",
  tidy_output = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the border or bidding zone}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{contract_type}{Contract market agreement type, valid values
can be checked from contract_types table;
"A01" = Day ahead
"A02" = Weekly
"A03" = Monthly
"A04" = Yearly
"A06" = Long Term
"A07" = Intraday
"A08" = Quarterly
Defaults to "A01" (Day ahead)}

\item{tidy_output}{Defaults to FALSE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Congestion income from implicit and flow-based allocations.
For implicit allocations, In_Domain and Out_Domain must be the same
border EIC code. For flow-based, both must be the same bidding zone EIC.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::congestion_income(
  eic = "10YDOM-1001A083J",
  period_start = lubridate::ymd(x = "2016-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2016-01-02", tz = "CET"),
  contract_type = "A01",
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other market endpoints:
\code{\link[=aggregated_bids]{aggregated_bids()}},
\code{\link[=allocated_transfer_capacities_3rd_countries]{allocated_transfer_capacities_3rd_countries()}},
\code{\link[=already_allocated_total_capacity]{already_allocated_total_capacity()}},
\code{\link[=auction_revenue]{auction_revenue()}},
\code{\link[=continuous_offered_transfer_capacities]{continuous_offered_transfer_capacities()}},
\code{\link[=continuous_offered_transfer_capacity]{continuous_offered_transfer_capacity()}},
\code{\link[=energy_prices]{energy_prices()}},
\code{\link[=explicit_offered_transfer_capacities]{explicit_offered_transfer_capacities()}},
\code{\link[=explicit_offered_transfer_capacity]{explicit_offered_transfer_capacity()}},
\code{\link[=flow_based_allocations]{flow_based_allocations()}},
\code{\link[=implicit_offered_transfer_capacities]{implicit_offered_transfer_capacities()}},
\code{\link[=implicit_offered_transfer_capacity]{implicit_offered_transfer_capacity()}},
\code{\link[=intraday_prices]{intraday_prices()}},
\code{\link[=net_positions]{net_positions()}},
\code{\link[=total_nominated_capacity]{total_nominated_capacity()}}
}
\concept{market endpoints}
