% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{costs_of_congestion_management}
\alias{costs_of_congestion_management}
\title{Get Costs of Congestion Management (13.1.C)}
\usage{
costs_of_congestion_management(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 31L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  event_nature = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{event_nature}{"A46" for system Operator redispatching
"B03" for counter trade
"B04" for congestion costs
Defaults to NULL which means both of them.}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Costs of TSO for redispatching and counter trading together
with costs for any other remedial actions taken to relieve
congested lines in transmission grid.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
# Belgium's Costs of Congestion Management
df <- entsoeapi::costs_of_congestion_management(
  eic = "10YBE----------2",
  period_start = lubridate::ymd(x = "2016-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2016-12-31", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other transmission endpoints:
\code{\link[=countertrading]{countertrading()}},
\code{\link[=cross_border_physical_flows]{cross_border_physical_flows()}},
\code{\link[=day_ahead_commercial_sched]{day_ahead_commercial_sched()}},
\code{\link[=expansion_and_dismantling_project]{expansion_and_dismantling_project()}},
\code{\link[=forecasted_transfer_capacities]{forecasted_transfer_capacities()}},
\code{\link[=intraday_cross_border_transfer_limits]{intraday_cross_border_transfer_limits()}},
\code{\link[=net_transfer_capacities]{net_transfer_capacities()}},
\code{\link[=redispatching_cross_border]{redispatching_cross_border()}},
\code{\link[=redispatching_internal]{redispatching_internal()}},
\code{\link[=total_commercial_sched]{total_commercial_sched()}}
}
\concept{transmission endpoints}
