% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_activated_balancing.R
\name{en_activated_balancing_reserves}
\alias{en_activated_balancing_reserves}
\title{Get Accepted Offers and Activated Balancing Reserves (4.6.11.) from Entso-e}
\usage{
en_activated_balancing_reserves(
  eic,
  period_start = lubridate::ymd(Sys.Date() - 1L, tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  reserve_type = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format  One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format  One year range limit applies}

\item{reserve_type}{Defaults to NULL, otherwise list of generation type codes (A95, A96, A97, A98) from StandardBusinessTypeList table}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Get Accepted Offers and Activated Balancing Reserves (4.6.11.) from Entso-e
}
\examples{

library(tidyverse)
library(entsoeapi)

de_lu_bal_202002 <- en_activated_balancing_reserves(eic          = "10Y1001A1001A82H",
                                                    period_start = lubridate::ymd("2020-02-01", tz = "CET"),
                                                    period_end   = lubridate::ymd("2020-03-01", tz = "CET"))

}
