% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_capacity.R
\name{en_capacity_total_nominated}
\alias{en_capacity_total_nominated}
\title{Get total nominated capacity from Entsoe}
\usage{
en_capacity_total_nominated(
  in_domain,
  out_domain,
  period_start,
  period_end,
  tidy_output = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT"),
  document_type = "A26",
  business_type = "B08"
)
}
\arguments{
\item{in_domain}{Energy Identification Code of the bidding zone or control area (TSO)}

\item{out_domain}{Energy Identification Code of the bidding zone or control area (TSO)}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format  One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format  One year range limit applies}

\item{tidy_output}{if TRUE then column names harmonized and definitive columns added, defaults to FALSE}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Get total nominated capacity from Entsoe
}
\examples{

library(tidyverse)
library(entsoeapi)

CZ_DE50HTZ_201902 <- en_capacity_total_nominated(in_domain = "10YDE-VE-------2",
                                                 out_domain = "10YCZ-CEPS-----N",
                                                 period_start = lubridate::ymd("2019-02-01", tz = "CET"),
                                                 period_end = lubridate::ymd("2019-03-01", tz = "CET"))

}
