% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{en_generation_act_gen_per_unit}
\alias{en_generation_act_gen_per_unit}
\title{Get actual generation output per generation unit from Entso-e (16.1.A)}
\usage{
en_generation_act_gen_per_unit(
  eic,
  period_start = lubridate::ymd(Sys.Date() - 1L, tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  gen_type = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area or bidding zone}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format  One day range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format  One day range limit applies}

\item{gen_type}{Defaults to NULL, otherwise list of generation type codes from StandardAssetTypeList table}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Get actual generation output per generation unit from Entso-e (16.1.A)
}
\examples{

library(tidyverse)
library(entsoeapi)

de_50Hertz_20200131 <- en_generation_act_gen_per_unit(eic = "10YDE-VE-------2",
                                                      period_start = lubridate::ymd("2020-01-31", tz = "CET"),
                                                      period_end = lubridate::ymd("2020-02-01", tz = "CET"),
                                                      gen_type = NULL,
                                                      tidy_output = TRUE)
View(de_50Hertz_20200131)

}
