% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{en_generation_agg_gen_per_type}
\alias{en_generation_agg_gen_per_type}
\title{Get total load from Entsoe}
\usage{
en_generation_agg_gen_per_type(eic, year, security_token = NULL)
}
\arguments{
\item{eic}{Energy Identification Code}

\item{security_token}{Security token}

\item{period_start}{POSIXct}

\item{period_end}{POSIXct}
}
\description{
Get total load from Entsoe
}
\examples{

library(tidyverse)
library(entsoeapi)

fr_2020 <- en_generation_agg_gen_per_type(eic = "10YFR-RTE------C", year = 2020)

}
