% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{en_generation_day_ahead_agg_gen}
\alias{en_generation_day_ahead_agg_gen}
\title{Get Day-ahead aggregated generation forecast from Entsoe.
It is an estimate of the total scheduled generation per bidding zone for the following day}
\usage{
en_generation_day_ahead_agg_gen(
  eic,
  period_start,
  period_end,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code}

\item{period_start}{POSIXct}

\item{period_end}{POSIXct}

\item{security_token}{Security token}
}
\description{
Get Day-ahead aggregated generation forecast from Entsoe.
It is an estimate of the total scheduled generation per bidding zone for the following day
}
\examples{

library(tidyverse)
library(entsoeapi)

fr_2020 <- en_generation_day_ahead_agg_gen(eic = "10YFR-RTE------C", period_start = lubridate::ymd("2020-02-01", tz = "CET"), period_end = lubridate::ymd("2020-03-01", tz = "CET"))

}
