% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{en_generation_day_ahead_gen_forecast_ws}
\alias{en_generation_day_ahead_gen_forecast_ws}
\title{Get total load from Entsoe}
\usage{
en_generation_day_ahead_gen_forecast_ws(
  eic,
  period_start,
  period_end,
  security_token = NULL
)
}
\arguments{
\item{eic}{Energy Identification Code}

\item{period_start}{POSIXct}

\item{period_end}{POSIXct}

\item{security_token}{Security token}
}
\description{
Get total load from Entsoe
}
\examples{

library(tidyverse)
library(entsoeapi)

fr_2020 <- en_generation_day_ahead_gen_forcast_ws(eic = "10YFR-RTE------C", period_start = lubridate::ymd("2020-02-01", tz = "CET"), period_end = lubridate::ymd("2020-03-01", tz = "CET"))

}
