% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_load_day_ahead_total_load_forecast.R
\name{en_load_day_ahead_total_load_forecast}
\alias{en_load_day_ahead_total_load_forecast}
\title{Get Day ahead total load forecast from Entsoe}
\usage{
en_load_day_ahead_total_load_forecast(eic, period_start, period_end,
  security_token = NULL)
}
\arguments{
\item{eic}{Energy Identification Code}

\item{period_start}{POSIXct}

\item{period_end}{POSIXct}

\item{security_token}{Security token}
}
\description{
Get Day ahead total load forecast from Entsoe
}
\examples{

library(tidyverse)
library(entsoeapi)
en_load_day_ahead_total_load_forecast(eic = "10Y1001A1001A82H", period_start = lubridate::ymd("2019-03-31", tz = "CET"), period_end = lubridate::ymd_hm("2019-03-31 23:00", tz = "CET"))

}
