% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages_generation_units.R
\name{en_outages}
\alias{en_outages}
\title{Get outages for Generation units.}
\usage{
en_outages(
  eic,
  mkt_doc_mrid = NULL,
  period_start = lubridate::ymd(Sys.Date(), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date() + 3, tz = "CET"),
  period_start_update = NULL,
  period_end_update = NULL,
  doc_status = NULL,
  business_type = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the bidding zone/control area (To extract outages of bidding zone  DE-AT-LU area, it is recommended to send queries per control area i.e. CTA|DE(50Hertz), CTA|DE(Amprion), CTA|DE(TeneTGer),CTA|DE(TransnetBW),CTA|AT,CTA|LU but not per bidding zone. )}

\item{mkt_doc_mrid}{Defaults to NULL. The older versions of an outage is returned only when mRID parameter is set.}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_start_update}{Period start update, POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end_update}{Period end update, POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{doc_status}{Defaults to NULL. Document status. "A05" for active, "A09" for cancelled and "A13" for withdrawn.}

\item{business_type}{Defaults to NULL. "A53" for planned maintenance. "A54" for unplanned outage.}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Get outages for Generation units.
}
\examples{

 library(tidyverse)
 library(entsoeapi)

 france <- en_outages(eic = "10YFR-RTE------C", period_start = lubridate::ymd("2019-11-12", tz = "CET"), period_end = lubridate::ymd("2019-11-13", tz = "CET"))

}
