% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages_generation_units.R
\name{en_outages_production_units}
\alias{en_outages_production_units}
\title{Get outages for production units.}
\usage{
en_outages_production_units(eic,
  period_start = lubridate::ymd(Sys.Date(), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date() + 3, tz = "CET"),
  doc_status = "A05", tidy_output = TRUE, security_token = NULL)
}
\arguments{
\item{eic}{Energy Identification Code}

\item{period_start}{POSIXct}

\item{period_end}{POSIXct}

\item{doc_status}{Document status. A05 for active or A09 for Cancelled.}

\item{security_token}{Security token}
}
\description{
Get outages for production units.
}
\examples{

 library(tidyverse)
 library(entsoeapi)

 france <- en_outages_production_units(eic = "10YFR-RTE------C", period_start = lubridate::ymd("2019-11-12", tz = "CET"), period_end = lubridate::ymd("2019-11-13", tz = "CET"))

}
