% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages_generation_units.R
\name{en_outages_transmission_infrastructure}
\alias{en_outages_transmission_infrastructure}
\title{Get outages for transmission units.}
\usage{
en_outages_transmission_infrastructure(
  in_domain,
  out_domain,
  period_start = lubridate::ymd(Sys.Date(), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date() + 3, tz = "CET"),
  period_start_update = NULL,
  period_end_update = NULL,
  doc_status = "A05",
  business_type = NULL,
  tidy_output = TRUE,
  new_style = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{period_start}{POSIXct}

\item{period_end}{POSIXct}

\item{period_start_update}{Period start udpate POSIXct}

\item{period_end_update}{Period end update POSIXct}

\item{doc_status}{Document status. "A05" for active, "A09" for cancelled and "A13" for withdrawn.}

\item{business_type}{Defaults to NULL. "A53" for planned maintenance. "A54" for unplanned outage.}

\item{tidy_output}{flatten nested tables}

\item{new_style}{If TRUE, some less useless descriptive columns are replaced by more useful ones}

\item{security_token}{Security token for ENTSO-E transparency platform}

\item{eic}{Energy Identification Code}
}
\description{
Get outages for transmission units.
}
\examples{

 library(tidyverse)
 library(entsoeapi)

 fr_de <- df <- en_outages_transmission_infrastructure(in_domain = "10YFR-RTE------C", out_domain = "10Y1001A1001A82H", period_start = lubridate::ymd(Sys.Date() - 300, tz = "CET"))

}
