% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission_forecasted_capacity.R
\name{en_transmission_day_ahead_commercial_schedules}
\alias{en_transmission_day_ahead_commercial_schedules}
\title{Get scheduled day-ahead commercial exchanges in aggregated form between bidding zones per direction and market time unit from Entsoe.}
\usage{
en_transmission_day_ahead_commercial_schedules(
  eic_in,
  eic_out,
  period_start,
  period_end,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of in domain}

\item{eic_out}{Energy Identification Code of out domain}

\item{period_start}{POSIXct}

\item{period_end}{POSIXct}

\item{security_token}{Security token}
}
\description{
Get scheduled day-ahead commercial exchanges in aggregated form between bidding zones per direction and market time unit from Entsoe.
}
\examples{

 library(tidyverse)
 library(entsoeapi)

 en_transmission_day_ahead_commercial_schedules(eic_in = "10YCZ-CEPS-----N", eic_out = "10YSK-SEPS-----K", period_start = lubridate::ymd("2019-11-01", tz = "CET"), period_end = lubridate::ymd("2019-12-01", tz = "CET"))
 en_transmission_day_ahead_commercial_schedules(eic_in = "10YDK-1--------W", eic_out = "10Y1001A1001A82H", period_start = lubridate::ymd("2019-11-01", tz = "CET"), period_end = lubridate::ymd("2019-12-01", tz = "CET"))

}
