% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{en_transmission_day_ahead_commercial_schedules_eic}
\alias{en_transmission_day_ahead_commercial_schedules_eic}
\title{Get the day-ahead commercial schedules EIC codes}
\usage{
en_transmission_day_ahead_commercial_schedules_eic()
}
\description{
Get the day-ahead commercial schedules EIC codes
}
