% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission_forecasted_capacity.R
\name{en_transmission_day_ahead_prices_eic}
\alias{en_transmission_day_ahead_prices_eic}
\title{Get the available areas.}
\usage{
en_transmission_day_ahead_prices_eic()
}
\description{
Get the available areas.
}
