% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission_x_border_phys_flow.R
\name{en_transmission_x_border_phys_flow}
\alias{en_transmission_x_border_phys_flow}
\title{Get cross-border physical flow values between control areas / bidding zones /countries from Entsoe.
Unlike Web GUI, API responds not netted values since data is requested per direction.}
\usage{
en_transmission_x_border_phys_flow(
  eic_in,
  eic_out,
  period_start,
  period_end,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of in domain}

\item{eic_out}{Energy Identification Code of out domain}

\item{period_start}{POSIXct}

\item{period_end}{POSIXct Minimum time interval in query response is an MTU period, but 1 year range limit applies.}

\item{security_token}{Security token}
}
\description{
Get cross-border physical flow values between control areas / bidding zones /countries from Entsoe.
Unlike Web GUI, API responds not netted values since data is requested per direction.
}
\examples{

 library(tidyverse)
 library(entsoeapi)

 en_transmission_x_border_phys_flow(eic_in = "10Y1001A1001A83F", eic_out = "10YCZ-CEPS-----N", period_start = lubridate::ymd("2020-01-01", tz = "CET"), period_end = lubridate::ymd("2020-01-02", tz = "CET"))
 en_transmission_x_border_phys_flow(eic_in = "10YCZ-CEPS-----N", eic_out = "10Y1001A1001A83F", period_start = lubridate::ymd("2020-01-01", tz = "CET"), period_end = lubridate::ymd("2020-01-02", tz = "CET"))

}
