% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entsoe_countries}
\alias{entsoe_countries}
\title{List of Mapped Countries for DAM Prices}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 43 rows and 2 columns.
}
\usage{
entsoe_countries
}
\description{
List of Mapped Countries for DAM Prices
}
\examples{

library(entsoeapi)

entsoe_countries

}
\keyword{datasets}
