% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoeapi-package.R
\name{entsoeapi}
\alias{entsoeapi}
\alias{entsoeapi-package}
\title{An R Interface to the ENTSO-E Transparency Platform API}
\description{
an easy wrapper around the ENTSO-E
\href{https://documenter.getpostman.com/view/7009892/2s93JtP3F6}{API}'s data and
transform them to tabular format without effort.
}
\keyword{internal}
