% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{explicit_offered_transfer_capacity}
\alias{explicit_offered_transfer_capacity}
\title{Get Explicit Offered Transfer Capacity (11.1.A)}
\usage{
explicit_offered_transfer_capacity(
  eic_in = NULL,
  eic_out = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  contract_type = "A01",
  tidy_output = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{eic_out}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{contract_type}{Contract market agreement type, valid values
can be checked from contract_types table;
A01 = Day ahead
A02 = Weekly
A03 = Monthly
A04 = Yearly
A06 = Long Term
A07 = Intraday
A08 = Quarterly
Defaults to "A01" (Day ahead)}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Explicit offered transfer capacity values.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::explicit_offered_transfer_capacity(
  eic_in = "10YBE----------2",
  eic_out = "10YGB----------A",
  period_start = lubridate::ymd(x = "2023-08-16", tz = "CET"),
  period_end = lubridate::ymd(x = "2023-08-17", tz = "CET"),
  contract_type = "A01",
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other market endpoints:
\code{\link[=aggregated_bids]{aggregated_bids()}},
\code{\link[=allocated_transfer_capacities_3rd_countries]{allocated_transfer_capacities_3rd_countries()}},
\code{\link[=already_allocated_total_capacity]{already_allocated_total_capacity()}},
\code{\link[=auction_revenue]{auction_revenue()}},
\code{\link[=congestion_income]{congestion_income()}},
\code{\link[=continuous_offered_transfer_capacities]{continuous_offered_transfer_capacities()}},
\code{\link[=continuous_offered_transfer_capacity]{continuous_offered_transfer_capacity()}},
\code{\link[=energy_prices]{energy_prices()}},
\code{\link[=explicit_offered_transfer_capacities]{explicit_offered_transfer_capacities()}},
\code{\link[=flow_based_allocations]{flow_based_allocations()}},
\code{\link[=implicit_offered_transfer_capacities]{implicit_offered_transfer_capacities()}},
\code{\link[=implicit_offered_transfer_capacity]{implicit_offered_transfer_capacity()}},
\code{\link[=intraday_prices]{intraday_prices()}},
\code{\link[=net_positions]{net_positions()}},
\code{\link[=total_nominated_capacity]{total_nominated_capacity()}}
}
\concept{market endpoints}
