% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{financial_expenses_and_income_for_balancing}
\alias{financial_expenses_and_income_for_balancing}
\title{Get Financial Expenses and Income for Balancing (17.1.I)}
\usage{
financial_expenses_and_income_for_balancing(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Financial expenses and income for balancing of the control area.
One year range limit applies.
}
\examples{
df <- entsoeapi::financial_expenses_and_income_for_balancing(
  eic = "10YCZ-CEPS-----N",
  period_start = lubridate::ymd(x = "2022-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2022-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
