% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{gen_day_ahead}
\alias{gen_day_ahead}
\title{Get Day-Ahead Generation Forecast. (14.1.C)}
\usage{
gen_day_ahead(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area,
bidding zone or country}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
It is an estimate of the total scheduled net generation (MW)
per area and market time unit of the following day.
}
\examples{

library(entsoeapi)
library(lubridate)

df <- gen_day_ahead(eic          = "10YFR-RTE------C",
                    period_start = ymd(x = "2020-02-01", tz = "CET"),
                    period_end   = ymd(x = "2020-03-01", tz = "CET"),
                    tidy_output  = TRUE)
str(df)

}
