% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{gen_installed_capacity_per_pt}
\alias{gen_installed_capacity_per_pt}
\title{Get Installed Generation Capacity per Production Type (14.1.A)}
\usage{
gen_installed_capacity_per_pt(
  eic = NULL,
  psr_type = NULL,
  year = year(Sys.Date()),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area,
bidding zone or country}

\item{psr_type}{Defaults to NULL, otherwise list of generation type
codes from asset_types table}

\item{year}{YYYY format}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
The sum of installed net generation capacity (MW) for all
existing production units equal to or exceeding 1 MW
installed generation capacity, per production type.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::gen_installed_capacity_per_pt(
  eic = "10YFR-RTE------C",
  psr_type = "B05",
  year = 2020
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other generation endpoints:
\code{\link[=gen_day_ahead_forecast]{gen_day_ahead_forecast()}},
\code{\link[=gen_installed_capacity_per_pu]{gen_installed_capacity_per_pu()}},
\code{\link[=gen_per_gen_unit]{gen_per_gen_unit()}},
\code{\link[=gen_per_prod_type]{gen_per_prod_type()}},
\code{\link[=gen_storage_mean_filling_rate]{gen_storage_mean_filling_rate()}},
\code{\link[=gen_wind_solar_forecasts]{gen_wind_solar_forecasts()}}
}
\concept{generation endpoints}
