% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{gen_per_gen_unit}
\alias{gen_per_gen_unit}
\title{Get Aggregated Generation per Generation Unit (16.1.A)}
\usage{
gen_per_gen_unit(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  gen_type = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area
or bidding zone}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{gen_type}{Defaults to NULL, otherwise list of generation type
codes from asset_types table}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Actual net generation output (MW) and optionally
consumption data from all generation units.
Data are aggregated as an average of generation outputs
or consumption.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::gen_per_gen_unit(
  eic          = "10YDE-VE-------2",
  period_start = lubridate::ymd(x = "2020-01-31", tz = "CET"),
  period_end   = lubridate::ymd(x = "2020-02-06", tz = "CET"),
  gen_type     = c("B04", "B05"),
  tidy_output  = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other generation endpoints:
\code{\link[=gen_day_ahead_forecast]{gen_day_ahead_forecast()}},
\code{\link[=gen_installed_capacity_per_pt]{gen_installed_capacity_per_pt()}},
\code{\link[=gen_installed_capacity_per_pu]{gen_installed_capacity_per_pu()}},
\code{\link[=gen_per_prod_type]{gen_per_prod_type()}},
\code{\link[=gen_storage_mean_filling_rate]{gen_storage_mean_filling_rate()}},
\code{\link[=gen_wind_solar_forecasts]{gen_wind_solar_forecasts()}}
}
\concept{generation endpoints}
