% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{gen_storage_mean_filling_rate}
\alias{gen_storage_mean_filling_rate}
\title{Get Weekly Average Filling Rate of Water Reservoirs
and Hydro Storage Plants (16.1.D)}
\usage{
gen_storage_mean_filling_rate(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 7L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area,
bidding zone or country}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
Maximum 380 days range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
Maximum 380 days range limit applies}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Aggregated weekly average filling rate of all water reservoir
and hydro storage plants (MWh) per area, including the same
week value of the previous year.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::gen_storage_mean_filling_rate(
  eic          = "10YFR-RTE------C",
  period_start = lubridate::ymd(x = "2020-01-01", tz = "CET"),
  period_end   = lubridate::ymd(x = "2020-01-31", tz = "CET"),
  tidy_output  = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other generation endpoints:
\code{\link[=gen_day_ahead_forecast]{gen_day_ahead_forecast()}},
\code{\link[=gen_installed_capacity_per_pt]{gen_installed_capacity_per_pt()}},
\code{\link[=gen_installed_capacity_per_pu]{gen_installed_capacity_per_pu()}},
\code{\link[=gen_per_gen_unit]{gen_per_gen_unit()}},
\code{\link[=gen_per_prod_type]{gen_per_prod_type()}},
\code{\link[=gen_wind_solar_forecasts]{gen_wind_solar_forecasts()}}
}
\concept{generation endpoints}
