% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{intraday_prices}
\alias{intraday_prices}
\title{Get Intraday Prices (12.1.D)}
\usage{
intraday_prices(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the related domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Prices in currency/MWh from the intraday market.
The data is delivered for each market time unit.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::intraday_prices(
  eic = "10YCZ-CEPS-----N",
  period_start = lubridate::ymd(x = "2024-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-02-01", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other market endpoints:
\code{\link[=aggregated_bids]{aggregated_bids()}},
\code{\link[=allocated_transfer_capacities_3rd_countries]{allocated_transfer_capacities_3rd_countries()}},
\code{\link[=already_allocated_total_capacity]{already_allocated_total_capacity()}},
\code{\link[=auction_revenue]{auction_revenue()}},
\code{\link[=congestion_income]{congestion_income()}},
\code{\link[=continuous_offered_transfer_capacities]{continuous_offered_transfer_capacities()}},
\code{\link[=continuous_offered_transfer_capacity]{continuous_offered_transfer_capacity()}},
\code{\link[=energy_prices]{energy_prices()}},
\code{\link[=explicit_offered_transfer_capacities]{explicit_offered_transfer_capacities()}},
\code{\link[=explicit_offered_transfer_capacity]{explicit_offered_transfer_capacity()}},
\code{\link[=flow_based_allocations]{flow_based_allocations()}},
\code{\link[=implicit_offered_transfer_capacities]{implicit_offered_transfer_capacities()}},
\code{\link[=implicit_offered_transfer_capacity]{implicit_offered_transfer_capacity()}},
\code{\link[=net_positions]{net_positions()}},
\code{\link[=total_nominated_capacity]{total_nominated_capacity()}}
}
\concept{market endpoints}
