% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_load.R
\name{load_day_ahead_total_forecast}
\alias{load_day_ahead_total_forecast}
\title{Get Day-Ahead Total Load Forecast (6.1.B)}
\usage{
load_day_ahead_total_forecast(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the bidding zone/
country/control area}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
It is defined as equal to the day-ahead forecasted sum of
power generated by plants on both TSO/DSO networks,
from which is deduced:
\itemize{
\item the balance (export-import) of exchanges on interconnections
between neighbouring bidding zones
\item the power absorbed by energy storage resources
}
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
# German average daily load.
df <- entsoeapi::load_day_ahead_total_forecast(
  eic = "10Y1001A1001A83F",
  period_start = lubridate::ymd(x = Sys.Date() -
    lubridate::days(x = 30), tz = "CET"),
  period_end = lubridate::ymd(x = Sys.Date(), tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other load endpoints:
\code{\link[=load_actual_total]{load_actual_total()}},
\code{\link[=load_month_ahead_total_forecast]{load_month_ahead_total_forecast()}},
\code{\link[=load_week_ahead_total_forecast]{load_week_ahead_total_forecast()}},
\code{\link[=load_year_ahead_forecast_margin]{load_year_ahead_forecast_margin()}},
\code{\link[=load_year_ahead_total_forecast]{load_year_ahead_total_forecast()}}
}
\concept{load endpoints}
