% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{m}
\alias{m}
\title{instantiate a memory cache store for maximum 1 hour}
\format{
An object of class \code{cache_mem} (inherits from \code{cachem}) of length 9.
}
\usage{
m
}
\description{
instantiate a memory cache store for maximum 1 hour
}
\keyword{datasets}
