% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{message_types}
\alias{message_types}
\title{List message types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 150 rows and 3 columns.
}
\usage{
message_types
}
\description{
List message types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
head(entsoeapi::message_types)

}
\seealso{
Other types:
\code{\link{allocation_mode_types}},
\code{\link{analog_types}},
\code{\link{area_types}},
\code{\link{asset_types}},
\code{\link{auction_types}},
\code{\link{business_types}},
\code{\link{category_types}},
\code{\link{classification_types}},
\code{\link{coding_scheme_types}},
\code{\link{connection_category_types}},
\code{\link{contract_types}},
\code{\link{coordinate_system_types}},
\code{\link{currency_types}},
\code{\link{curve_types}},
\code{\link{customer_types}},
\code{\link{direction_types}},
\code{\link{document_types}},
\code{\link{eic_types}},
\code{\link{energy_product_types}},
\code{\link{flow_commodity_option_types}},
\code{\link{fuel_types}},
\code{\link{grid_agreement_types}},
\code{\link{hvdc_mode_types}},
\code{\link{indicator_types}},
\code{\link{market_product_types}},
\code{\link{object_aggregation_types}},
\code{\link{payment_terms_types}},
\code{\link{price_category_types}},
\code{\link{price_component_types}},
\code{\link{price_direction_types}},
\code{\link{process_types}},
\code{\link{quality_types}},
\code{\link{reason_code_types}},
\code{\link{rights_types}},
\code{\link{role_types}},
\code{\link{settlement_method_types}},
\code{\link{status_types}},
\code{\link{sub_area_types}},
\code{\link{tariff_types}},
\code{\link{timeframe_types}},
\code{\link{unit_multiplier}},
\code{\link{unit_of_measure_types}},
\code{\link{unit_symbol_types}}
}
\concept{types}
\keyword{datasets}
