% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages.R
\name{outages_transmission_grid}
\alias{outages_transmission_grid}
\title{Get Unavailability of Transmission Infrastructure. (10.1.A&B)}
\usage{
outages_transmission_grid(
  eic_in = NULL,
  eic_out = NULL,
  period_start = ymd(Sys.Date() + days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date() + days(x = 3L), tz = "CET"),
  period_start_update = NULL,
  period_end_update = NULL,
  doc_status = NULL,
  event_nature = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of the IN bidding zone area}

\item{eic_out}{Energy Identification Code of the OUT bidding zone area}

\item{period_start}{the starting date of the in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{the ending date of the outage in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_start_update}{notification submission/update starting date
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end_update}{notification submission/update ending date
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{doc_status}{Notification document status. "A05" for active,
"A09" for cancelled and "A13" for withdrawn.
Defaults to NULL which means "A05" and "A09" together.}

\item{event_nature}{"A53" for planned maintenance.
"A54" for unplanned outage.
Defaults to NULL which means both of them.}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
The planned and forced unavailability, including changes in
unavailability of interconnections in the transmission grid
that reduce transfer capacities between areas during at least
one market time unit including information about new
net transfer capacity.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::outages_transmission_grid(
  eic_in = "10YFR-RTE------C",
  eic_out = "10Y1001A1001A82H",
  period_start = lubridate::ymd(
    x = Sys.Date() + lubridate::days(x = 1),
    tz = "CET"
  ),
  period_end = lubridate::ymd(
    x = Sys.Date() + lubridate::days(x = 2),
    tz = "CET"
  ),
  period_start_update = lubridate::ymd(
    x = Sys.Date() - lubridate::days(x = 7),
    tz = "CET"
  ),
  period_end_update = lubridate::ymd(x = Sys.Date(), tz = "CET")
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other outage endpoints:
\code{\link[=outages_both]{outages_both()}},
\code{\link[=outages_cons_units]{outages_cons_units()}},
\code{\link[=outages_fallbacks]{outages_fallbacks()}},
\code{\link[=outages_gen_units]{outages_gen_units()}},
\code{\link[=outages_offshore_grid]{outages_offshore_grid()}},
\code{\link[=outages_prod_units]{outages_prod_units()}}
}
\concept{outage endpoints}
