% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{party_eic}
\alias{party_eic}
\title{Get Party_X Energy Identification Codes}
\usage{
party_eic()
}
\value{
A tibble of accordingly filtered EIC codes, which contains such columns as
\code{eic_code}, \code{eic_display_name}, \code{eic_long_name}, \code{eic_parent},
\code{eic_responsible_party}, \code{eic_status}, \code{market_participant_postal_code},
\code{market_participant_iso_country_code}, \code{market_participant_vat_code},
\code{eic_type_function_list} and \code{type}.
}
\description{
This function downloads approved party X
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
It covers market participants.
}
\examples{
\dontshow{if (there_is_provider()) withAutoprint(\{ # examplesIf}
eic_party <- entsoeapi::party_eic()

dplyr::glimpse(eic_party)
\dontshow{\}) # examplesIf}
}
