% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{redispatching_internal}
\alias{redispatching_internal}
\title{Get Redispatching Internal (13.1.A)}
\usage{
redispatching_internal(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Changes in production and load (increase or decrease) to
relieve internal congestion lines that exceeds its capacity.
100 documents limit applies!!
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
# Netherlands' internal redispatching.
df <- entsoeapi::redispatching_internal(
  eic = "10YNL----------L",
  period_start = lubridate::ymd(x = "2023-11-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2023-12-01", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other transmission endpoints:
\code{\link[=costs_of_congestion_management]{costs_of_congestion_management()}},
\code{\link[=countertrading]{countertrading()}},
\code{\link[=cross_border_physical_flows]{cross_border_physical_flows()}},
\code{\link[=day_ahead_commercial_sched]{day_ahead_commercial_sched()}},
\code{\link[=expansion_and_dismantling_project]{expansion_and_dismantling_project()}},
\code{\link[=forecasted_transfer_capacities]{forecasted_transfer_capacities()}},
\code{\link[=intraday_cross_border_transfer_limits]{intraday_cross_border_transfer_limits()}},
\code{\link[=net_transfer_capacities]{net_transfer_capacities()}},
\code{\link[=redispatching_cross_border]{redispatching_cross_border()}},
\code{\link[=total_commercial_sched]{total_commercial_sched()}}
}
\concept{transmission endpoints}
