% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_congestion.R
\name{redispatching_x_border}
\alias{redispatching_x_border}
\title{title
Get Redispatching Cross Border (13.1.A)}
\usage{
redispatching_x_border(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of the control area}

\item{eic_out}{Energy Identification Code of the control area}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Changes in production and load (increase or decrease) to
relieve congested internal lines that exceeds its capacity.
100 documents limit applies!!
}
\examples{

library(entsoeapi)
library(lubridate)

# Germany's cross-border redispatching between TenneT and 50Hertz TSO.
df <- redispatching_x_border(
  eic_in       = "10YDE-EON------1",
  eic_out      = "10YDE-VE-------2",
  period_start = ymd(x = "2024-09-01", tz = "CET"),
  period_end   = ymd(x = "2024-10-01", tz = "CET"),
  tidy_output  = TRUE
)

str(df)

}
