% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{resource_object_eic}
\alias{resource_object_eic}
\title{Get Resource Object W Energy Identification Codes}
\usage{
resource_object_eic()
}
\value{
A tibble of accordingly filtered EIC codes, which contains such columns as
\code{eic_code}, \code{eic_display_name}, \code{eic_long_name}, \code{eic_parent},
\code{eic_responsible_party}, \code{eic_status}, \code{market_participant_postal_code},
\code{market_participant_iso_country_code}, \code{market_participant_vat_code},
\code{eic_type_function_list} and \code{type}.
}
\description{
This function downloads approved resource object W
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
A resource that can either produce or consume energy
and that is reported in a schedule.
}
\examples{
\dontshow{if (there_is_provider()) withAutoprint(\{ # examplesIf}
eic_resource_object <- entsoeapi::resource_object_eic()

dplyr::glimpse(eic_resource_object)
\dontshow{\}) # examplesIf}
}
