% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{rr_actual_capacity}
\alias{rr_actual_capacity}
\title{Get RR Actual Capacity (SO GL 189.3)}
\usage{
rr_actual_capacity(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the area domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Actual capacity of Replacement Reserve (RR).
Process type is hardcoded to A46 (Replacement Reserve).
}
\examples{
df <- entsoeapi::rr_actual_capacity(
  eic = "10YAT-APG------L",
  period_start = lubridate::ymd(x = "2022-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2022-04-01", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
