% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{sharing_of_rr_and_frr_capacity}
\alias{sharing_of_rr_and_frr_capacity}
\title{Get Sharing of RR and FRR Capacity (SO GL 190.1)}
\usage{
sharing_of_rr_and_frr_capacity(
  eic_acquiring = NULL,
  eic_connecting = NULL,
  process_type = NULL,
  period_start = ymd(Sys.Date() - days(x = 7L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_acquiring}{Energy Identification Code of the acquiring domain}

\item{eic_connecting}{Energy Identification Code of the connecting domain}

\item{process_type}{type of reserve
"A56" FRR
"A46" RR}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Sharing of Replacement Reserve and Frequency Restoration Reserve capacity
in the same Synchronous Area.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
\dontrun{
df1 <- entsoeapi::sharing_of_rr_and_frr_capacity(
  eic_acquiring = "10YCB-GERMANY--8",
  eic_connecting = "10YAT-APG------L",
  process_type = "A46",
  period_start = lubridate::ymd(x = "2025-10-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2025-12-31", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df1)

df2 <- entsoeapi::sharing_of_rr_and_frr_capacity(
  eic_acquiring = "10YCB-GERMANY--8",
  eic_connecting = "10YAT-APG------L",
  process_type = "A56",
  period_start = lubridate::ymd(x = "2025-10-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2025-12-31", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df2)
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other balancing endpoints:
\code{\link[=activated_balancing_prices]{activated_balancing_prices()}},
\code{\link[=aggregated_balancing_energy_bids]{aggregated_balancing_energy_bids()}},
\code{\link[=allocation_of_cross_zonal_balancing_cap]{allocation_of_cross_zonal_balancing_cap()}},
\code{\link[=balancing_border_cap_limit]{balancing_border_cap_limit()}},
\code{\link[=balancing_energy_bids]{balancing_energy_bids()}},
\code{\link[=changes_to_bid_availability]{changes_to_bid_availability()}},
\code{\link[=contracted_reserves]{contracted_reserves()}},
\code{\link[=current_balancing_state]{current_balancing_state()}},
\code{\link[=elastic_demands]{elastic_demands()}},
\code{\link[=exchanged_volumes]{exchanged_volumes()}},
\code{\link[=exchanged_volumes_per_border]{exchanged_volumes_per_border()}},
\code{\link[=fcr_total_capacity]{fcr_total_capacity()}},
\code{\link[=financial_expenses_and_income]{financial_expenses_and_income()}},
\code{\link[=hvdc_link_constrains]{hvdc_link_constrains()}},
\code{\link[=imbalance_prices]{imbalance_prices()}},
\code{\link[=imbalance_volumes]{imbalance_volumes()}},
\code{\link[=netted_volumes]{netted_volumes()}},
\code{\link[=netted_volumes_per_border]{netted_volumes_per_border()}},
\code{\link[=procured_balancing_capacity]{procured_balancing_capacity()}},
\code{\link[=rr_and_frr_actual_capacity]{rr_and_frr_actual_capacity()}},
\code{\link[=shares_of_fcr_capacity]{shares_of_fcr_capacity()}}
}
\concept{balancing endpoints}
