% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardAssetTypeList}
\alias{StandardAssetTypeList}
\title{StandardAssetTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 29 rows and 3 columns.
}
\usage{
StandardAssetTypeList
}
\description{
StandardAssetTypeList
}
\keyword{datasets}
