% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardBusinessTypeList}
\alias{StandardBusinessTypeList}
\title{StandardBusinessTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 133 rows and 3 columns.
}
\usage{
StandardBusinessTypeList
}
\description{
StandardBusinessTypeList
}
\keyword{datasets}
