% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardCodingSchemeTypeList}
\alias{StandardCodingSchemeTypeList}
\title{StandardCodingSchemeTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 48 rows and 3 columns.
}
\usage{
StandardCodingSchemeTypeList
}
\description{
StandardCodingSchemeTypeList
}
\keyword{datasets}
