% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardContractTypeList}
\alias{StandardContractTypeList}
\title{StandardContractTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 12 rows and 3 columns.
}
\usage{
StandardContractTypeList
}
\description{
StandardContractTypeList
}
\keyword{datasets}
