% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardCurrencyTypeList}
\alias{StandardCurrencyTypeList}
\title{StandardCurrencyTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 22 rows and 3 columns.
}
\usage{
StandardCurrencyTypeList
}
\description{
StandardCurrencyTypeList
}
\keyword{datasets}
