% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardCurveTypeList}
\alias{StandardCurveTypeList}
\title{StandardCurveTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 5 rows and 3 columns.
}
\usage{
StandardCurveTypeList
}
\description{
StandardCurveTypeList
}
\keyword{datasets}
