% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardDirectionTypeList}
\alias{StandardDirectionTypeList}
\title{StandardDirectionTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
StandardDirectionTypeList
}
\description{
StandardDirectionTypeList
}
\keyword{datasets}
