% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardDocumentTypeList}
\alias{StandardDocumentTypeList}
\title{StandardDocumentTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 103 rows and 3 columns.
}
\usage{
StandardDocumentTypeList
}
\description{
StandardDocumentTypeList
}
\keyword{datasets}
