% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardEnergyProductTypeList}
\alias{StandardEnergyProductTypeList}
\title{StandardEnergyProductTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 7 rows and 3 columns.
}
\usage{
StandardEnergyProductTypeList
}
\description{
StandardEnergyProductTypeList
}
\keyword{datasets}
