% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardHVDCModeTypeList}
\alias{StandardHVDCModeTypeList}
\title{StandardHVDCModeTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
StandardHVDCModeTypeList
}
\description{
StandardHVDCModeTypeList
}
\keyword{datasets}
