% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardObjectAggregationTypeList}
\alias{StandardObjectAggregationTypeList}
\title{StandardObjectAggregationTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11 rows and 3 columns.
}
\usage{
StandardObjectAggregationTypeList
}
\description{
StandardObjectAggregationTypeList
}
\keyword{datasets}
