% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardPriceDirectionTypeList}
\alias{StandardPriceDirectionTypeList}
\title{StandardPriceDirectionTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2 rows and 3 columns.
}
\usage{
StandardPriceDirectionTypeList
}
\description{
StandardPriceDirectionTypeList
}
\keyword{datasets}
