% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardProcessTypeList}
\alias{StandardProcessTypeList}
\title{StandardProcessTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 42 rows and 3 columns.
}
\usage{
StandardProcessTypeList
}
\description{
StandardProcessTypeList
}
\keyword{datasets}
