% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardReasonCodeTypeList}
\alias{StandardReasonCodeTypeList}
\title{StandardReasonCodeTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 106 rows and 3 columns.
}
\usage{
StandardReasonCodeTypeList
}
\description{
StandardReasonCodeTypeList
}
\keyword{datasets}
