% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardRightsTypeList}
\alias{StandardRightsTypeList}
\title{StandardRightsTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 6 rows and 3 columns.
}
\usage{
StandardRightsTypeList
}
\description{
StandardRightsTypeList
}
\keyword{datasets}
