% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardRoleTypeList}
\alias{StandardRoleTypeList}
\title{StandardRoleTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 41 rows and 3 columns.
}
\usage{
StandardRoleTypeList
}
\description{
StandardRoleTypeList
}
\keyword{datasets}
