% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardStatusTypeList}
\alias{StandardStatusTypeList}
\title{StandardStatusTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 16 rows and 3 columns.
}
\usage{
StandardStatusTypeList
}
\description{
StandardStatusTypeList
}
\keyword{datasets}
