% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StandardUnitOfMeasureTypeList}
\alias{StandardUnitOfMeasureTypeList}
\title{StandardUnitOfMeasureTypeList}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 14 rows and 3 columns.
}
\usage{
StandardUnitOfMeasureTypeList
}
\description{
StandardUnitOfMeasureTypeList
}
\keyword{datasets}
