% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{substation_eic}
\alias{substation_eic}
\title{Get Substation_A EIC codes}
\usage{
substation_eic()
}
\description{
This function downloads all approved Energy Identification Codes from
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes/
}
\examples{
eic_substation <- substation_eic()
}
