% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{tie_line_eic}
\alias{tie_line_eic}
\title{Get Tie Line T Energy Identification Codes}
\usage{
tie_line_eic()
}
\value{
A tibble of accordingly filtered EIC codes, which contains such columns as
\code{eic_code}, \code{eic_display_name}, \code{eic_long_name}, \code{eic_parent},
\code{eic_responsible_party}, \code{eic_status}, \code{market_participant_postal_code},
\code{market_participant_iso_country_code}, \code{market_participant_vat_code},
\code{eic_type_function_list} and \code{type}.
}
\description{
This function downloads approved tie line T
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
It covers a transmission line that connects different areas
excluding HVDC interconnectors.
}
\examples{
\dontshow{if (there_is_provider()) withAutoprint(\{ # examplesIf}
eic_tie_line <- entsoeapi::tie_line_eic()

dplyr::glimpse(eic_tie_line)
\dontshow{\}) # examplesIf}
}
