% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{total_commercial_sched}
\alias{total_commercial_sched}
\title{Get Total Commercial Schedules (12.1.F)}
\usage{
total_commercial_sched(
  eic_in = NULL,
  eic_out = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of in domain}

\item{eic_out}{Energy Identification Code of out domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Aggregated capacity nominated for all time horizons
(including Intra-Day) corresponding to implicit and
explicit allocations after each nomination process.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df1 <- entsoeapi::total_commercial_sched(
  eic_in = "10YCZ-CEPS-----N",
  eic_out = "10YSK-SEPS-----K",
  period_start = lubridate::ymd(x = "2019-11-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2019-12-01", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df1)

df2 <- entsoeapi::total_commercial_sched(
  eic_in = "10YDK-1--------W",
  eic_out = "10Y1001A1001A82H",
  period_start = lubridate::ymd(x = "2019-11-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2019-12-01", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other transmission endpoints:
\code{\link[=costs_of_congestion_management]{costs_of_congestion_management()}},
\code{\link[=countertrading]{countertrading()}},
\code{\link[=cross_border_physical_flows]{cross_border_physical_flows()}},
\code{\link[=day_ahead_commercial_sched]{day_ahead_commercial_sched()}},
\code{\link[=expansion_and_dismantling_project]{expansion_and_dismantling_project()}},
\code{\link[=forecasted_transfer_capacities]{forecasted_transfer_capacities()}},
\code{\link[=intraday_cross_border_transfer_limits]{intraday_cross_border_transfer_limits()}},
\code{\link[=net_transfer_capacities]{net_transfer_capacities()}},
\code{\link[=redispatching_cross_border]{redispatching_cross_border()}},
\code{\link[=redispatching_internal]{redispatching_internal()}}
}
\concept{transmission endpoints}
