% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{transm_already_allocated_cap}
\alias{transm_already_allocated_cap}
\title{Get Total Already Allocated Capacity (12.1.C)}
\usage{
transm_already_allocated_cap(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  auction_type = "A04",
  contract_type = "A05",
  tidy_output = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{eic_out}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{auction_type}{Auction category, valid values can be checked
from auction_types table;
Defaults to "A04" (Mixed)}

\item{contract_type}{Contract market agreement type, valid values
can be checked from contract_types table;
Defaults to "A05" (Total)}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Total capacity allocated, for all time horizons
(including Intra-Day) after each allocation process
per market time unit.
}
\examples{
df <- entsoeapi::transm_already_allocated_cap(
  eic_in        = "10YDE-VE-------2",
  eic_out       = "10YCZ-CEPS-----N",
  period_start  = lubridate::ymd(x = "2019-02-01", tz = "CET"),
  period_end    = lubridate::ymd(x = "2019-02-02", tz = "CET"),
  auction_type  = "A02",
  contract_type = "A01"
)
str(df)

}
