% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{transm_day_ahead_comm_sched}
\alias{transm_day_ahead_comm_sched}
\title{Get Day-Ahead Commercial Schedules (12.1.F)}
\usage{
transm_day_ahead_comm_sched(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of in domain}

\item{eic_out}{Energy Identification Code of out domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Day-ahead commercial exchanges in aggregated form between
bidding zones per direction and market time unit.
}
\examples{
df1 <- entsoeapi::transm_day_ahead_comm_sched(
  eic_in       = "10YCZ-CEPS-----N",
  eic_out      = "10YSK-SEPS-----K",
  period_start = lubridate::ymd(x = "2019-11-01", tz = "CET"),
  period_end   = lubridate::ymd(x = "2019-12-01", tz = "CET"),
  tidy_output  = TRUE
)
str(df1)

df2 <- entsoeapi::transm_day_ahead_comm_sched(
  eic_in       = "10YDK-1--------W",
  eic_out      = "10Y1001A1001A82H",
  period_start = lubridate::ymd(x = "2019-11-01", tz = "CET"),
  period_end   = lubridate::ymd(x = "2019-12-01", tz = "CET"),
  tidy_output  = TRUE
)
str(df2)

}
