% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{transm_day_ahead_prices}
\alias{transm_day_ahead_prices}
\title{Get Day-Ahead Prices (12.1.D)}
\usage{
transm_day_ahead_prices(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the related domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Prices in currency/MWh created on spot (Day-Ahead) market.
The data is delivered for each market time unit.
}
\examples{
df1 <- entsoeapi::transm_day_ahead_prices(
  eic          = "10YCZ-CEPS-----N",
  period_start = lubridate::ymd(x = "2019-11-01", tz = "CET"),
  period_end   = lubridate::ymd(x = "2019-12-01", tz = "CET"),
  tidy_output  = TRUE
)
str(df1)

df2 <- entsoeapi::transm_day_ahead_prices(
  eic          = "10YDK-1--------W",
  period_start = lubridate::ymd(x = "2019-11-01", tz = "CET"),
  period_end   = lubridate::ymd(x = "2019-12-01", tz = "CET"),
  tidy_output  = TRUE
)
str(df2)

}
