% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{transm_total_nominated_cap}
\alias{transm_total_nominated_cap}
\title{Get Total Nominated Capacity (12.1.B)}
\usage{
transm_total_nominated_cap(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of in domain}

\item{eic_out}{Energy Identification Code of out domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Aggregated capacity nominated by market participants from
time horizons (including Intra-Day) corresponding to explicit
allocations, agreed between the TSOs and confirmed
to the market.
}
\examples{
df <- entsoeapi::transm_total_nominated_cap(
  eic_in       = "10YDE-VE-------2",
  eic_out      = "10YCZ-CEPS-----N",
  period_start = lubridate::ymd(x = "2019-02-01", tz = "CET"),
  period_end   = lubridate::ymd(x = "2019-03-01", tz = "CET"),
  tidy_output  = TRUE
)

str(df)

}
