% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{transm_x_border_phys_flow}
\alias{transm_x_border_phys_flow}
\title{Get Cross-Border Physical Flow (12.1.G)}
\usage{
transm_x_border_phys_flow(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of in domain}

\item{eic_out}{Energy Identification Code of out domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
Minimum time interval in query response is an MTU period,
but 1 year range limit applies.}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
It is the measured real flow of energy between
the neighbouring areas on the cross borders.
}
\examples{

library(entsoeapi)
library(lubridate)

df1 <- transm_x_border_phys_flow(
  eic_in       = "10Y1001A1001A83F",
  eic_out      = "10YCZ-CEPS-----N",
  period_start = ymd(x = "2020-01-01", tz = "CET"),
  period_end   = ymd(x = "2020-01-02", tz = "CET"),
  tidy_output  = TRUE
)

str(df1)

df2 <- transm_x_border_phys_flow(
  eic_in       = "10YCZ-CEPS-----N",
  eic_out      = "10Y1001A1001A83F",
  period_start = ymd(x = "2020-01-01", tz = "CET"),
  period_end   = ymd(x = "2020-01-02", tz = "CET"),
  tidy_output  = TRUE
)

str(df2)

}
