% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transmission_pair_eic_dict}
\alias{transmission_pair_eic_dict}
\title{List transmission pair EIC Dictionary}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 114 rows and 8 columns.
}
\usage{
transmission_pair_eic_dict
}
\description{
List transmission pair EIC Dictionary
}
\examples{
dplyr::glimpse(entsoeapi::transmission_pair_eic_dict)

}
\keyword{datasets}
