% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{accounting_point_eic}
\alias{accounting_point_eic}
\title{Get Accounting_point_Z EIC codes}
\usage{
accounting_point_eic()
}
\description{
This function downloads approved Accounting_point_Z Energy Identification Codes from
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes/
}
\examples{
eic_accounting_point <- accounting_point_eic()
}
