% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{activated_balancing_prices}
\alias{activated_balancing_prices}
\title{Get Activated Balancing Prices (TR 17.1.F, IF aFRR 3.16)}
\usage{
activated_balancing_prices(
  eic = NULL,
  process_type = "A16",
  business_type = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area domain}

\item{process_type}{Process type code, defaults to "A16"}

\item{business_type}{Optional business type code}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Prices of activated balancing energy and aFRR cross-border marginal prices.
One year range limit applies.
}
\examples{
df <- entsoeapi::activated_balancing_prices(
  eic = "10YCZ-CEPS-----N",
  period_start = lubridate::ymd(x = "2024-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
