% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{aggregated_balancing_energy_bids}
\alias{aggregated_balancing_energy_bids}
\title{Get Aggregated Balancing Energy Bids (GL EB 12.3.E)}
\usage{
aggregated_balancing_energy_bids(
  eic = NULL,
  process_type = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the area domain}

\item{process_type}{type of frequency restoration reserve
"A51" aFRR
"A46" RR
"A47" mFRR
"A60" mFRR with scheduled activation
"A61" mFRR with direct activation}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Aggregated balancing energy bids.
One year range limit applies.
}
\examples{
df <- entsoeapi::aggregated_balancing_energy_bids(
  eic = "10YCZ-CEPS-----N",
  process_type = "A51",
  period_start = lubridate::ymd(x = "2024-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
