% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{aggregated_bids}
\alias{aggregated_bids}
\title{Get Aggregated Bids (12.3.D)}
\usage{
aggregated_bids(
  eic = NULL,
  process_type = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the area domain}

\item{process_type}{Process type code:
"A51" = aFRR
"A46" = Replacement Reserve
"A47" = Manual Frequency Restoration Reserve
"A60" = mFRR with scheduled activation
"A61" = mFRR with direct activation
"A67" = Frequency Containment Reserve
"A68" = Automatic and Manual Frequency
Restoration Reserve}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Aggregated filled and unfilled bids per market time unit.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::aggregated_bids(
  eic = "10YCZ-CEPS-----N",
  process_type = "A51",
  period_start = lubridate::ymd(x = "2024-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-01-02", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other market endpoints: 
\code{\link{allocated_transfer_capacities_3rd_countries}()},
\code{\link{already_allocated_total_capacity}()},
\code{\link{auction_revenue}()},
\code{\link{congestion_income}()},
\code{\link{continuous_offered_transfer_capacities}()},
\code{\link{continuous_offered_transfer_capacity}()},
\code{\link{energy_prices}()},
\code{\link{explicit_offered_transfer_capacities}()},
\code{\link{explicit_offered_transfer_capacity}()},
\code{\link{flow_based_allocations}()},
\code{\link{implicit_offered_transfer_capacities}()},
\code{\link{implicit_offered_transfer_capacity}()},
\code{\link{intraday_prices}()},
\code{\link{net_positions}()},
\code{\link{total_nominated_capacity}()}
}
\concept{market endpoints}
