% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{all_approved_eic}
\alias{all_approved_eic}
\title{Get all Approved Energy Identification Codes}
\usage{
all_approved_eic()
}
\description{
This function downloads approved all
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
Further details are under:
https://www.entsoe.eu/data/energy-identification-codes-eic/#eic-documentation
}
\examples{

library(entsoeapi)

eic_all <- all_approved_eic()

str(eic_all)

}
