% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{all_approved_eic}
\alias{all_approved_eic}
\title{Get all Approved Energy Identification Codes}
\usage{
all_approved_eic()
}
\value{
A tibble of accordingly filtered EIC codes, which contains such columns as
\code{EicCode}, \code{EicDisplayName}, \code{EicLongName}, \code{EicParent},
\code{EicResponsibleParty}, \code{EicStatus}, \code{MarketParticipantPostalCode},
\code{MarketParticipantIsoCountryCode}, \code{MarketParticipantVatCode},
\code{EicTypeFunctionList} and \code{type}.
}
\description{
This function downloads all approved
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
Further details are under:
https://www.entsoe.eu/data/energy-identification-codes-eic/#eic-documentation
}
\examples{
eic_all <- entsoeapi::all_approved_eic()

str(eic_all)

}
