% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{allocated_transfer_capacities_3rd_countries}
\alias{allocated_transfer_capacities_3rd_countries}
\title{Get Transfer Capacities Allocated with Third Countries — Explicit (12.1.H)}
\usage{
allocated_transfer_capacities_3rd_countries(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  contract_type = "A01",
  auction_category = "A04",
  position = 1L,
  tidy_output = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of the in domain}

\item{eic_out}{Energy Identification Code of the out domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{contract_type}{Contract market agreement type, valid values
can be checked from contract_types table;
"A01" = Day ahead
"A02" = Weekly
"A03" = Monthly
"A04" = Yearly
"A06" = Long Term
"A07" = Intraday
"A08" = Quarterly
Defaults to "A01" (Day ahead)}

\item{auction_category}{Optional auction category;
"A01" = Base
"A02" = Peak
"A03" = Off Peak
"A04" = Hourly
Defaults to "A04"}

\item{position}{Integer position for ts_classification_sequence_position.
Defaults to 1.}

\item{tidy_output}{Defaults to FALSE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Capacity allocated outside the EU via explicit auction (auction.Type=A02).
Minimum time interval in query response ranges from part of day to year,
depending on the selected contract type. A 100-document limit applies.
}
\examples{
df <- entsoeapi::allocated_transfer_capacities_3rd_countries(
  eic_in = "10YSK-SEPS-----K",
  eic_out = "10YUA-WEPS-----0",
  period_start = lubridate::ymd(x = "2016-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2016-01-02", tz = "CET"),
  contract_type = "A01",
  auction_category = "A04",
  position = 1L,
  tidy_output = TRUE
)

str(df)

}
