% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allocation_mode_types}
\alias{allocation_mode_types}
\title{List Allocation Mode Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
allocation_mode_types
}
\description{
List Allocation Mode Types
}
\examples{

str(entsoeapi::allocation_mode_types)

}
\keyword{datasets}
