% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{allocation_of_cross_zonal_balancing_cap}
\alias{allocation_of_cross_zonal_balancing_cap}
\title{Get Allocation of Cross-Zonal Balancing Capacity (GL EB 12.3.H&I)}
\usage{
allocation_of_cross_zonal_balancing_cap(
  eic_acquiring = NULL,
  eic_connecting = NULL,
  market_agreement_type = NULL,
  period_start = ymd(Sys.Date() - days(x = 7L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_acquiring}{Energy Identification Code of the acquiring domain}

\item{eic_connecting}{Energy Identification Code of the connecting domain}

\item{market_agreement_type}{Optional market agreement type code
"A01" Daily
"A02" Weekly
"A06" Long term}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Use of allocated cross-zonal balancing capacity between
an acquiring and a connecting domain.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
\dontrun{
df <- entsoeapi::allocation_of_cross_zonal_balancing_cap(
  eic_acquiring = "10YAT-APG------L",
  eic_connecting = "10YCH-SWISSGRIDZ",
  period_start = lubridate::ymd(x = "2024-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-01-02", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other balancing endpoints: 
\code{\link{activated_balancing_prices}()},
\code{\link{aggregated_balancing_energy_bids}()},
\code{\link{balancing_border_cap_limit}()},
\code{\link{balancing_energy_bids}()},
\code{\link{changes_to_bid_availability}()},
\code{\link{contracted_reserves}()},
\code{\link{current_balancing_state}()},
\code{\link{elastic_demands}()},
\code{\link{exchanged_volumes}()},
\code{\link{exchanged_volumes_per_border}()},
\code{\link{fcr_total_capacity}()},
\code{\link{financial_expenses_and_income}()},
\code{\link{hvdc_link_constrains}()},
\code{\link{imbalance_prices}()},
\code{\link{imbalance_volumes}()},
\code{\link{netted_volumes}()},
\code{\link{netted_volumes_per_border}()},
\code{\link{procured_balancing_capacity}()},
\code{\link{rr_and_frr_actual_capacity}()},
\code{\link{shares_of_fcr_capacity}()},
\code{\link{sharing_of_rr_and_frr_capacity}()}
}
\concept{balancing endpoints}
