% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{allocation_of_cross_zonal_balancing_cap}
\alias{allocation_of_cross_zonal_balancing_cap}
\title{Get Allocation of Cross-Zonal Balancing Capacity (GL EB 12.3.H&I)}
\usage{
allocation_of_cross_zonal_balancing_cap(
  eic_acquiring = NULL,
  eic_connecting = NULL,
  type_market_agreement = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_acquiring}{Energy Identification Code of the acquiring domain}

\item{eic_connecting}{Energy Identification Code of the connecting domain}

\item{type_market_agreement}{Optional market agreement type code}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Use of allocated cross-zonal balancing capacity between
an acquiring and a connecting domain.
}
\examples{
df <- entsoeapi::allocation_of_cross_zonal_balancing_cap(
  eic_acquiring = "10YAT-APG------L",
  eic_connecting = "10YCH-SWISSGRIDZ",
  period_start = lubridate::ymd(x = "2024-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
