% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{already_allocated_total_capacity}
\alias{already_allocated_total_capacity}
\title{Get Already Allocated Total Capacity (12.1.C)}
\usage{
already_allocated_total_capacity(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  auction_category = "A04",
  contract_type = "A01",
  tidy_output = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{eic_out}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{auction_category}{A01 = Base
A02 = Peak
A03 = Off Peak
A04 = Hourly
Defaults to "A04" (Hourly)}

\item{contract_type}{Contract market agreement type, valid values
can be checked from contract_types table;
A01 = Daily
A02 = Weekly
A03 = Monthly
A04 = Yearly
A06 = Long Term
A07 = Intraday
A08 = Quarterly
Defaults to "A01" (Daily)}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Total capacity allocated, for all time horizons
(including Intra-Day) after each allocation process
per market time unit.
}
\examples{
df <- entsoeapi::already_allocated_total_capacity(
  eic_in = "10YDE-VE-------2",
  eic_out = "10YCZ-CEPS-----N",
  period_start = lubridate::ymd(x = "2025-02-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2025-02-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
