% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analog_types}
\alias{analog_types}
\title{List Analog Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 42 rows and 3 columns.
}
\usage{
analog_types
}
\description{
List Analog Types
}
\examples{

library(entsoeapi)

str(analog_types)

}
\keyword{datasets}
