% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{area_eic}
\alias{area_eic}
\title{Get Area_Y EIC codes}
\usage{
area_eic()
}
\description{
This function downloads approved Area_Y Energy Identification Codes from
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes/
}
\examples{
eic_area <- area_eic()
}
