% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{area_eic}
\alias{area_eic}
\title{Get Area Y Energy Identification Codes}
\usage{
area_eic()
}
\value{
A tibble of accordingly filtered EIC codes, which contains such columns as
\code{EicCode}, \code{EicDisplayName}, \code{EicLongName}, \code{EicParent}, \code{EicResponsibleParty},
\code{EicStatus}, \code{MarketParticipantPostalCode}, \code{MarketParticipantIsoCountryCode},
\code{MarketParticipantVatCode}, \code{EicTypeFunctionList} and \code{type}.
}
\description{
This function downloads approved area Y
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
}
\examples{

library(entsoeapi)

eic_area <- area_eic()

str(eic_area)

}
