% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_types}
\alias{asset_types}
\title{List Asset Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 90 rows and 3 columns.
}
\usage{
asset_types
}
\description{
List Asset Types
}
\examples{

library(entsoeapi)

str(asset_types)

}
\keyword{datasets}
