% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auction_types}
\alias{auction_types}
\title{List Auction Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8 rows and 3 columns.
}
\usage{
auction_types
}
\description{
List Auction Types
}
\examples{

library(entsoeapi)

str(auction_types)

}
\keyword{datasets}
