% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{balancing_border_cap_limit}
\alias{balancing_border_cap_limit}
\title{Balancing Border Capacity Limitations (IFs 4.3 & 4.4)}
\usage{
balancing_border_cap_limit(
  eic_in = NULL,
  eic_out = NULL,
  process_type = "A51",
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of an in LFC Area (LFA)
or in Scheduling area (SCA)}

\item{eic_out}{Energy Identification Code of out an out LFC Area (LFA)
or out Scheduling area (SCA)}

\item{process_type}{type of frequency restoration reserve
"A47" mFRR
"A51" aFRR
"A63" Imbalance Netting}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data, or \code{NULL} if no data
is available for the given parameters.
}
\description{
This data item publish limitations on borders requested by
participating or affected TSOs.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::balancing_border_cap_limit(
  eic_in = "10YDE-RWENET---I",
  eic_out = "10YBE----------2",
  process_type = "A51",
  period_start = lubridate::ymd(x = "2022-06-22", tz = "CET"),
  period_end = lubridate::ymd(x = "2022-06-23", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
