% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{balancing_energy_bids}
\alias{balancing_energy_bids}
\title{Get Balancing Energy Bids (GL EB 12.3.B&C)}
\usage{
balancing_energy_bids(
  eic = NULL,
  process_type = "A47",
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the connecting domain}

\item{process_type}{Process type (currently only A47 is supported)}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Balancing energy bids submitted by BSPs.
100 documents limit applies; paging is handled transparently.
}
\examples{
df <- entsoeapi::balancing_energy_bids(
  eic = "10YCZ-CEPS-----N",
  process_type = "A47",
  period_start = lubridate::ymd(x = "2024-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
