% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{business_types}
\alias{business_types}
\title{List Business Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 292 rows and 3 columns.
}
\usage{
business_types
}
\description{
List Business Types
}
\examples{

library(entsoeapi)

str(business_types)

}
\keyword{datasets}
