% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{category_types}
\alias{category_types}
\title{List Category Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
category_types
}
\description{
List Category Types
}
\examples{

library(entsoeapi)

str(category_types)

}
\keyword{datasets}
