% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{changes_to_bid_availability}
\alias{changes_to_bid_availability}
\title{Get Changes to Bid Availability (IFs mFRR 9.9, aFRR 9.6 & 9.8)}
\usage{
changes_to_bid_availability(
  eic = NULL,
  business_type = "C46",
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the domain}

\item{business_type}{type of frequency restoration reserve
"C40" Conditional bid
"C41" Thermal limit
"C42" Frequency limit
"C43" Voltage limit
"C44" Current limit
"C45" Short-circuit current limits
"C46" Dynamic stability limit
Defaults to "C46".}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Changes to bid availability for scheduled activation products.
100 documents limit applies; paging is handled transparently.
}
\examples{
df <- entsoeapi::changes_to_bid_availability(
  eic = "10YCZ-CEPS-----N",
  business_type = "C46",
  period_start = lubridate::ymd(x = "2024-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
