% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classification_types}
\alias{classification_types}
\title{List Classification Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2 rows and 3 columns.
}
\usage{
classification_types
}
\description{
List Classification Types
}
\examples{

library(entsoeapi)

str(classification_types)

}
\keyword{datasets}
