% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coding_scheme_types}
\alias{coding_scheme_types}
\title{List coding scheme types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 51 rows and 3 columns.
}
\usage{
coding_scheme_types
}
\description{
List coding scheme types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
dplyr::glimpse(entsoeapi::coding_scheme_types)

}
\keyword{datasets}
