% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{congestion_income}
\alias{congestion_income}
\title{Get Implicit and Flow-based Allocations — Congestion Income (12.1.E)}
\usage{
congestion_income(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  contract_type = "A01",
  tidy_output = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the border or bidding zone}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{contract_type}{Contract market agreement type, valid values
can be checked from contract_types table;
"A01" = Day ahead
"A02" = Weekly
"A032 = Monthly
"A04" = Yearly
"A06" = Long Term
"A07" = Intraday
"A08" = Quarterly
Defaults to "A01" (Day ahead)}

\item{tidy_output}{Defaults to FALSE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Congestion income from implicit and flow-based allocations.
For implicit allocations, In_Domain and Out_Domain must be the same
border EIC code. For flow-based, both must be the same bidding zone EIC.
}
\examples{
df <- entsoeapi::congestion_income(
  eic = "10YDOM-1001A083J",
  period_start = lubridate::ymd(x = "2016-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2016-01-02", tz = "CET"),
  contract_type = "A01",
  tidy_output = TRUE
)

str(df)

}
