% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{connection_category_types}
\alias{connection_category_types}
\title{List connection category types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
connection_category_types
}
\description{
List connection category types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
dplyr::glimpse(entsoeapi::connection_category_types)

}
\keyword{datasets}
