% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{continuous_offered_transfer_capacity}
\alias{continuous_offered_transfer_capacity}
\title{Get Continuous Offered Transfer Capacity (11.1)}
\usage{
continuous_offered_transfer_capacity(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{eic_out}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Continuous offered transfer capacity values.
}
\examples{
df <- entsoeapi::continuous_offered_transfer_capacity(
  eic_in = "10YNL----------L",
  eic_out = "10YBE----------2",
  period_start = lubridate::ymd(x = "2024-05-16", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-05-17", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
