% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{contracted_reserves}
\alias{contracted_reserves}
\title{Get Contracted Reserves (17.1.B&C)}
\usage{
contracted_reserves(
  eic = NULL,
  type_market_agreement = NULL,
  process_type = NULL,
  psr_type = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area domain}

\item{type_market_agreement}{Market agreement type code (mandatory)
"A01" Daily
"A02" Weekly
"A03" Monthly
"A04" Yearly
"A06" Long term
"A13" Hourly}

\item{process_type}{Optional process type code
"A46" Replacement reserve
"A47" Manual frequency restoration reserve
"A51" Automatic frequency restoration reserve
"A52" Frequency containment reserve}

\item{psr_type}{Optional PSR type code
"A03" Mixed
"A04" Generation
"A05" Load}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Volumes and prices of contracted reserves.
100 documents limit applies; paging is handled transparently.
}
\examples{
df <- entsoeapi::contracted_reserves(
  eic = "10YCZ-CEPS-----N",
  type_market_agreement = "A13",
  period_start = lubridate::ymd(x = "2022-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2022-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
