% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coordinate_system_types}
\alias{coordinate_system_types}
\title{List Coordinate System Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
coordinate_system_types
}
\description{
List Coordinate System Types
}
\examples{

str(entsoeapi::coordinate_system_types)

}
\keyword{datasets}
