% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_congestion.R
\name{costs_of_congestion_management}
\alias{costs_of_congestion_management}
\title{Get Costs of Congestion Management (13.1.C)}
\usage{
costs_of_congestion_management(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 31L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  event_nature = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{event_nature}{"A46" for system Operator redispatching
"B03" for counter trade
"B04" for congestion costs
Defaults to NULL which means both of them.}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Costs of TSO for redispatching and counter trading together
with costs for any other remedial actions taken to relieve
congested lines in transmission grid.
100 documents limit applies!!
}
\examples{

library(entsoeapi)
library(lubridate)

# Czech's Costs of Congestion Management
df <- costs_of_congestion_management(
  eic          = "10YCZ-CEPS-----N",
  period_start = ymd(x = "2016-01-01", tz = "CET"),
  period_end   = ymd(x = "2017-01-01", tz = "CET"),
  tidy_output  = TRUE
)

str(df)

}
