% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{countertrading}
\alias{countertrading}
\title{title
Get Countertrading (13.1.B)}
\usage{
countertrading(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of the control area/bidding zone}

\item{eic_out}{Energy Identification Code of the control area/bidding zone}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Buying or cancelling generation on different side of the border
to relieve congested cross-border lines that exceeds its capacity.
The time interval in the query response depends on duration of
matching counter trades
100 documents limit applies!!
}
\examples{
# Counter trading between Germany and Denmark.
df <- entsoeapi::countertrading(
  eic_in = "10Y1001A1001A82H",
  eic_out = "10YDK-1--------W",
  period_start = lubridate::ymd(x = "2024-09-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-10-01", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
