% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{currency_types}
\alias{currency_types}
\title{List Currency Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 22 rows and 3 columns.
}
\usage{
currency_types
}
\description{
List Currency Types
}
\examples{

library(entsoeapi)

str(currency_types)

}
\keyword{datasets}
