% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{current_balancing_state}
\alias{current_balancing_state}
\title{Get Current Balancing State (GL EB 12.3.A)}
\usage{
current_balancing_state(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the area domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
100 day range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
100 day range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Current balancing state of the control area.
100 day range limit applies.
}
\examples{
df <- entsoeapi::current_balancing_state(
  eic = "10YCZ-CEPS-----N",
  period_start = lubridate::ymd(x = "2024-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
