% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curve_types}
\alias{curve_types}
\title{List Curve Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 5 rows and 3 columns.
}
\usage{
curve_types
}
\description{
List Curve Types
}
\examples{

library(entsoeapi)

str(curve_types)

}
\keyword{datasets}
