% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customer_types}
\alias{customer_types}
\title{List customer types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
customer_types
}
\description{
List customer types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
dplyr::glimpse(entsoeapi::customer_types)

}
\keyword{datasets}
