% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{direction_types}
\alias{direction_types}
\title{List direction types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
direction_types
}
\description{
List direction types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
dplyr::glimpse(entsoeapi::direction_types)

}
\keyword{datasets}
