% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{direction_types}
\alias{direction_types}
\title{List Direction Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
direction_types
}
\description{
List Direction Types
}
\examples{

str(entsoeapi::direction_types)

}
\keyword{datasets}
