% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{document_types}
\alias{document_types}
\title{List Document Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 103 rows and 3 columns.
}
\usage{
document_types
}
\description{
List Document Types
}
\examples{

library(entsoeapi)

str(document_types)

}
\keyword{datasets}
