% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eic_types}
\alias{eic_types}
\title{List EIC Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 7 rows and 3 columns.
}
\usage{
eic_types
}
\description{
List EIC Types
}
\examples{

str(entsoeapi::eic_types)

}
\keyword{datasets}
