% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{en_generation_agg_gen_per_type}
\alias{en_generation_agg_gen_per_type}
\title{Get aggregated generation per type from Entsoe}
\usage{
en_generation_agg_gen_per_type(
  eic,
  period_start,
  period_end,
  security_token = Sys.getenv("ENTSOE_PAT"),
  gen_type = NULL
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area (TSO)}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format  One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format  One year range limit applies}

\item{security_token}{Security token for ENTSO-E transparency platform}

\item{gen_type}{Defaults to NULL, otherwise list of generation type codes from StandardAssetTypeList table}
}
\description{
Get aggregated generation per type from Entsoe
}
\examples{

library(tidyverse)
library(entsoeapi)

fr_202002 <- en_generation_agg_gen_per_type(eic = "10YFR-RTE------C",
                                            period_start = lubridate::ymd("2020-02-01", tz = "CET"),
                                            period_end = lubridate::ymd("2020-03-01", tz = "CET"))

}
