% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{en_generation_agg_gen_per_type}
\alias{en_generation_agg_gen_per_type}
\title{Get aggregated generation per type from Entso-e (16.1.B&C)}
\usage{
en_generation_agg_gen_per_type(
  eic,
  period_start = lubridate::ymd(Sys.Date() - 1L, tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  gen_type = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area, bidding zone or country}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format  One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format  One year range limit applies}

\item{gen_type}{Defaults to NULL, otherwise list of generation type codes from StandardAssetTypeList table}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Get aggregated generation per type from Entso-e (16.1.B&C)
}
\examples{

library(tidyverse)
library(entsoeapi)

fr_202002 <- en_generation_agg_gen_per_type(eic = "10YFR-RTE------C",
                                            period_start = lubridate::ymd("2020-02-01", tz = "CET"),
                                            period_end = lubridate::ymd("2020-03-01", tz = "CET"),
                                            gen_type       = NULL,
                                            tidy_output    = TRUE)
View(fr_202002)

}
