% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages_generation_units.R
\name{en_outages}
\alias{en_outages}
\title{Get outages for Generation units.}
\usage{
en_outages(
  eic,
  period_start = lubridate::ymd(Sys.Date(), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date() + 3, tz = "CET"),
  period_start_update = NULL,
  period_end_update = NULL,
  doc_status = "A05",
  tidy_output = TRUE,
  security_token = NULL
)
}
\arguments{
\item{eic}{Energy Identification Code}

\item{period_start}{POSIXct}

\item{period_end}{POSIXct}

\item{period_start_update}{Period start udpate.}

\item{period_end_update}{Period end update.}

\item{doc_status}{Document status. A05 for active or A09 for Cancelled.}

\item{security_token}{Security token}
}
\description{
Get outages for Generation units.
}
\examples{

 library(tidyverse)
 library(entsoeapi)

 france <- en_outages(eic = "10YFR-RTE------C", period_start = lubridate::ymd("2019-11-12", tz = "CET"), period_end = lubridate::ymd("2019-11-13", tz = "CET"))

}
