% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages_generation_units.R
\name{en_outages_clean}
\alias{en_outages_clean}
\title{Clean data for outages}
\usage{
en_outages_clean(out_df)
}
\arguments{
\item{out_df}{Outages data.frame}
}
\description{
Clean data for outages
}
