% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages_generation_units.R
\name{en_outages_transmission_infrastructure}
\alias{en_outages_transmission_infrastructure}
\title{Get outages for transmission units.}
\usage{
en_outages_transmission_infrastructure(
  in_domain,
  out_domain,
  period_start = lubridate::ymd(Sys.Date(), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date() + 3, tz = "CET"),
  period_start_update = NULL,
  period_end_update = NULL,
  doc_status = "A05",
  tidy_output = TRUE,
  security_token = NULL
)
}
\arguments{
\item{period_start}{POSIXct}

\item{period_end}{POSIXct}

\item{period_start_update}{Period start udpate.}

\item{period_end_update}{Period end update.}

\item{doc_status}{Document status. A05 for active or A09 for Cancelled.}

\item{security_token}{Security token}

\item{eic}{Energy Identification Code}
}
\description{
Get outages for transmission units.
}
\examples{

 library(tidyverse)
 library(entsoeapi)

 fr_de <- df <- en_outages_transmission_infrastructure(in_domain = "10YFR-RTE------C", out_domain = "10Y1001A1001A82H", period_start = lubridate::ymd(Sys.Date() - 300, tz = "CET"))

}
