% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission_forecasted_capacity.R
\name{en_transmission_day_ahead_commercial_schedules_eic}
\alias{en_transmission_day_ahead_commercial_schedules_eic}
\title{EIC codes}
\usage{
en_transmission_day_ahead_commercial_schedules_eic()
}
\description{
EIC codes
}
