% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{energy_prices}
\alias{energy_prices}
\title{Get Day-Ahead or Intraday Energy Prices (12.1.D)}
\usage{
energy_prices(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  contract_type = "A01",
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the related domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{contract_type}{Contract market agreement type;
"A01" = Day ahead
"A07" = Intraday
Defaults to "A01" (Day ahead)}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Prices in currency/MWh created on spot (Day-Ahead or Intraday) market.
The data is delivered for each market time unit.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df1 <- entsoeapi::energy_prices(
  eic = "10YCZ-CEPS-----N",
  period_start = lubridate::ymd(x = "2019-11-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2019-12-01", tz = "CET"),
  contract_type = "A01",
  tidy_output = TRUE
)

dplyr::glimpse(df1)

df2 <- entsoeapi::energy_prices(
  eic = "10YES-REE------0",
  period_start = lubridate::ymd(x = "2026-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2026-01-02", tz = "CET"),
  contract_type = "A07",
  tidy_output = TRUE
)

dplyr::glimpse(df2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other market endpoints: 
\code{\link{aggregated_bids}()},
\code{\link{allocated_transfer_capacities_3rd_countries}()},
\code{\link{already_allocated_total_capacity}()},
\code{\link{auction_revenue}()},
\code{\link{congestion_income}()},
\code{\link{continuous_offered_transfer_capacities}()},
\code{\link{continuous_offered_transfer_capacity}()},
\code{\link{explicit_offered_transfer_capacities}()},
\code{\link{explicit_offered_transfer_capacity}()},
\code{\link{flow_based_allocations}()},
\code{\link{implicit_offered_transfer_capacities}()},
\code{\link{implicit_offered_transfer_capacity}()},
\code{\link{intraday_prices}()},
\code{\link{net_positions}()},
\code{\link{total_nominated_capacity}()}
}
\concept{market endpoints}
