% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy_product_types}
\alias{energy_product_types}
\title{List Energy Product Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 9 rows and 3 columns.
}
\usage{
energy_product_types
}
\description{
List Energy Product Types
}
\examples{

library(entsoeapi)

str(energy_product_types)

}
\keyword{datasets}
