% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{exchanged_volumes_per_border}
\alias{exchanged_volumes_per_border}
\title{Get Netted & Exchanged Volumes Per Border (IFs 3.10, 3.16 & 3.17)}
\usage{
exchanged_volumes_per_border(
  acquiring_eic = NULL,
  connecting_eic = NULL,
  process_type = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{acquiring_eic}{Energy Identification Code of the acquiring area}

\item{connecting_eic}{Energy Identification Code of the connecting area}

\item{process_type}{type of frequency restoration reserve
"A51" aFRR
"A60" mFRR with scheduled activation
"A61" mFRR with direct activation}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One day range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One day range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Netted and exchanged balancing energy volumes on a specific border between
an acquiring and a connecting domain.
}
\examples{
df <- entsoeapi::exchanged_volumes_per_border(
  acquiring_eic = "10YBE----------2",
  connecting_eic = "10YFR-RTE------C",
  process_type = "A60",
  period_start = lubridate::ymd(x = "2025-03-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2025-03-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
