% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{expansion_and_dismantling_project}
\alias{expansion_and_dismantling_project}
\title{Get Expansion and Dismantling Projects (9.1)}
\usage{
expansion_and_dismantling_project(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  business_type = NULL,
  doc_status = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of in domain}

\item{eic_out}{Energy Identification Code of out domain}

\item{period_start}{the starting date of the in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{the ending date of the outage in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{business_type}{"B01" = for interconnector network evolution
"B02" = interconnector network dismantling}

\item{doc_status}{Notification document status.
"A01" for intermediate
"A02" for final
"A05" for active,
"A09" for cancelled
"A13" for withdrawn
"X01" for estimated}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A data frame with expansion or dismantling projects
}
\description{
Query the interconnector network evolution or dismantling
}
\examples{
df <- entsoeapi::expansion_and_dismantling_project(
  eic_in = "10YSK-SEPS-----K",
  eic_out = "10YHU-MAVIR----U",
  period_start = lubridate::ymd(x = "2023-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2023-01-02", tz = "CET"),
  business_type = "B01",
  doc_status = "A05",
  tidy_output = TRUE
)

str(df)

}
