% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{explicit_offered_transfer_capacity}
\alias{explicit_offered_transfer_capacity}
\title{Get Explicit Offered Transfer Capacity (11.1.A)}
\usage{
explicit_offered_transfer_capacity(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  contract_type = "A01",
  tidy_output = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{eic_out}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{contract_type}{Contract market agreement type, valid values
can be checked from contract_types table;
A01 = Day ahead
A02 = Weekly
A03 = Monthly
A04 = Yearly
A06 = Long Term
A07 = Intraday
A08 = Quarterly
Defaults to "A01" (Day ahead)}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Explicit offered transfer capacity values.
}
\examples{
df <- entsoeapi::explicit_offered_transfer_capacity(
  eic_in = "10YBE----------2",
  eic_out = "10YGB----------A",
  period_start = lubridate::ymd(x = "2023-08-16", tz = "CET"),
  period_end = lubridate::ymd(x = "2023-08-17", tz = "CET"),
  contract_type = "A01",
  tidy_output = TRUE
)

str(df)

}
