% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{fcr_total_capacity}
\alias{fcr_total_capacity}
\title{Get FCR Total Capacity (SO GL 187.2)}
\usage{
fcr_total_capacity(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the area domain}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Total Frequency Containment Reserve capacity.
One year range limit applies.
}
\examples{
df <- entsoeapi::fcr_total_capacity(
  eic = "10YEU-CONT-SYNC0",
  period_start = lubridate::ymd(x = "2022-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2022-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
