% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{flow_based_allocations}
\alias{flow_based_allocations}
\title{Get Flow Based Allocations (11.1.B)}
\usage{
flow_based_allocations(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  process_type = "A43",
  archive = FALSE,
  tidy_output = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the area}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{process_type}{Contract market agreement type, valid values
can be checked from process_types table;
"A32" = Month-ahead
"A33" = Year-ahead
"A43" = Day ahead
"A44" = Intraday
Defaults to "A43" (Day ahead)}

\item{archive}{Defaults to FALSE, set to TRUE if archives to be queried.}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Flow based capacity allocated, for all time horizons.
}
\examples{
df1 <- entsoeapi::flow_based_allocations(
  eic = "10Y1001A1001A91G",
  period_start = lubridate::ymd(x = "2025-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2026-01-01", tz = "CET"),
  process_type = "A43",
  archive = FALSE,
  tidy_output = TRUE
)

str(df1)

df2 <- entsoeapi::flow_based_allocations(
  eic = "10YDOM-REGION-1V",
  period_start = lubridate::ymd(x = "2018-12-31", tz = "CET"),
  period_end = lubridate::ymd(x = "2019-01-01", tz = "CET"),
  process_type = "A32",
  archive = TRUE,
  tidy_output = TRUE
)

str(df2)

}
