% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_commodity_option_types}
\alias{flow_commodity_option_types}
\title{List Flow Commodity Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
flow_commodity_option_types
}
\description{
List Flow Commodity Types
}
\examples{

str(entsoeapi::flow_commodity_option_types)

}
\keyword{datasets}
