% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuel_types}
\alias{fuel_types}
\title{List Fuel Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 55 rows and 3 columns.
}
\usage{
fuel_types
}
\description{
List Fuel Types
}
\examples{

str(entsoeapi::fuel_types)

}
\keyword{datasets}
