% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{gen_installed_capacity_per_pt}
\alias{gen_installed_capacity_per_pt}
\title{Get Installed Generation Capacity per Production Type (14.1.A)}
\usage{
gen_installed_capacity_per_pt(
  eic = NULL,
  psr_type = NULL,
  year = lubridate::year(Sys.Date()),
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area,
bidding zone or country}

\item{psr_type}{Defaults to NULL, otherwise list of generation type
codes from asset_types table}

\item{year}{YYYY format}

\item{security_token}{Security token}
}
\description{
The sum of installed net generation capacity (MW) for all
existing production units equalling to or exceeding 1 MW
installed generation capacity, per production type.
}
\examples{

library(entsoeapi)

df <- gen_installed_capacity_per_pt(eic      = "10YFR-RTE------C",
                                    psr_type = "B05",
                                    year     = 2020)
str(df)

}
