% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{gen_installed_capacity_per_pu}
\alias{gen_installed_capacity_per_pu}
\title{Get Installed Generation Capacity per Production Unit (14.1.B)}
\usage{
gen_installed_capacity_per_pu(
  eic = NULL,
  year = lubridate::year(Sys.Date()),
  psr_type = NULL,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area,
bidding zone or country}

\item{year}{YYYY format Cannot be shown more than 3 years ahead
as required by the law.}

\item{psr_type}{Defaults to NULL, otherwise list of generation type
codes from asset_types table}

\item{security_token}{Security token}
}
\description{
The installed generation capacities (MW) at the beginning of the year
for all the production units, including the planned ones.
}
\examples{
df <- entsoeapi::gen_installed_capacity_per_pu(
  eic      = "10YDE-VE-------2",
  year     = 2020,
  psr_type = "B05"
)
str(df)

}
