% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{gen_per_gen_unit}
\alias{gen_per_gen_unit}
\title{Get Aggregated Generation per Generation Unit (16.1.A)}
\usage{
gen_per_gen_unit(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  gen_type = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area
or bidding zone}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One day range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One day range limit applies}

\item{gen_type}{Defaults to NULL, otherwise list of generation type
codes from asset_types table}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Actual net generation output (MW) and optionally
consumption data from all generation units.
Data are aggregated as an average of generation outputs
or consumption.
}
\examples{

library(entsoeapi)
library(lubridate)

df <- gen_per_gen_unit(eic          = "10YDE-VE-------2",
                       period_start = ymd(x = "2020-01-31", tz = "CET"),
                       period_end   = ymd(x = "2020-02-01", tz = "CET"),
                       gen_type     = NULL,
                       tidy_output  = TRUE)
str(df)

}
