% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{gen_per_prod_type}
\alias{gen_per_prod_type}
\title{Get Aggregated Generation per Production Type (16.1.B&C)}
\usage{
gen_per_prod_type(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  gen_type = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area,
bidding zone or country}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{gen_type}{Defaults to NULL, otherwise list of generation type
codes from asset_types table}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Actual aggregated net generation output (MW) or consumption
per market time unit and per production type.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::gen_per_prod_type(
  eic          = "10YFR-RTE------C",
  period_start = lubridate::ymd(x = "2020-02-01", tz = "CET"),
  period_end   = lubridate::ymd(x = "2020-03-01", tz = "CET"),
  gen_type     = NULL,
  tidy_output  = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other generation endpoints: 
\code{\link{gen_day_ahead_forecast}()},
\code{\link{gen_installed_capacity_per_pt}()},
\code{\link{gen_installed_capacity_per_pu}()},
\code{\link{gen_per_gen_unit}()},
\code{\link{gen_storage_mean_filling_rate}()},
\code{\link{gen_wind_solar_forecasts}()}
}
\concept{generation endpoints}
