% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{gen_per_prod_type}
\alias{gen_per_prod_type}
\title{Get Aggregated Generation per Production Type (16.1.B&C)}
\usage{
gen_per_prod_type(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  gen_type = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area,
bidding zone or country}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
Maximum one year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
Maximum one year range limit applies}

\item{gen_type}{Defaults to NULL, otherwise list of generation type
codes from asset_types table}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Actual aggregated net generation output (MW) or consumption
per market time unit and per production type.
}
\examples{
df <- entsoeapi::gen_per_prod_type(
  eic          = "10YFR-RTE------C",
  period_start = lubridate::ymd(x = "2020-02-01", tz = "CET"),
  period_end   = lubridate::ymd(x = "2020-03-01", tz = "CET"),
  gen_type     = NULL,
  tidy_output  = TRUE
)
str(df)

}
