% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{gen_storage_mean_filling_rate}
\alias{gen_storage_mean_filling_rate}
\title{Get Weekly Average Filling Rate of Water Reservoirs
and Hydro Storage Plants (16.1.D)}
\usage{
gen_storage_mean_filling_rate(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area,
bidding zone or country}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
Maximum 380 days range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
Maximum 380 days range limit applies}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Aggregated weekly average filling rate of all water reservoir
and hydro storage plants (MWh) per area, including the same
week value of the previous year.
}
\examples{
df <- entsoeapi::gen_storage_mean_filling_rate(
  eic          = "10YFR-RTE------C",
  period_start = lubridate::ymd(x = "2020-02-01", tz = "CET"),
  period_end   = lubridate::ymd(x = "2021-02-15", tz = "CET"),
  tidy_output  = TRUE)
str(df)

}
