% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_generation.R
\name{gen_wind_solar_forecasts}
\alias{gen_wind_solar_forecasts}
\title{Get Generation Forecasts for Wind & Solar (14.1.D)}
\usage{
gen_wind_solar_forecasts(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date(), tz = "CET") - lubridate::days(x = 1L),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area,
bidding zone or country}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
A respective forecast of wind and solar power net
generation (MW) per area and each market time unit
of the following/current day.
The elements of the result list are representing
each related forecast time range.
}
\examples{

library(entsoeapi)
library(lubridate)

df_list <- gen_wind_solar_forecasts(eic          = "10YFR-RTE------C",
                                    period_start = ymd(x = "2020-02-01",
                                                       tz = "CET"),
                                    period_end   = ymd(x = "2020-03-01",
                                                       tz = "CET"),
                                    tidy_output  = TRUE)
str(df_list$`Day-ahead`)
str(df_list$`Intraday`)
str(df_list$`Current`)

}
