% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_agreement_types}
\alias{grid_agreement_types}
\title{List grid agreement types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
grid_agreement_types
}
\description{
List grid agreement types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
dplyr::glimpse(entsoeapi::grid_agreement_types)

}
\keyword{datasets}
