% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_agreement_types}
\alias{grid_agreement_types}
\title{List Grid Agreement Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
grid_agreement_types
}
\description{
List Grid Agreement Types
}
\examples{

str(entsoeapi::grid_agreement_types)

}
\keyword{datasets}
