% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{hvdc_link_constrains}
\alias{hvdc_link_constrains}
\title{Get HVDC Link Constraints (IFs 4.5)}
\usage{
hvdc_link_constrains(
  eic_in = NULL,
  eic_out = NULL,
  eic_ic = NULL,
  process_type = "A63",
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of the in domain}

\item{eic_out}{Energy Identification Code of the out domain}

\item{eic_ic}{Energy Identification Code of the interconnector
If used, data for the given Transmission Asset is returned}

\item{process_type}{"A47" Manual frequency restoration reserve
"A51" Automatic Frequency Restoration Reserve
"A63" Imbalance Netting;
Defaults to "A63"}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Permanent allocation limitations to cross-border capacity on HVDC lines.
}
\examples{
df <- entsoeapi::hvdc_link_constrains(
  eic_in = "10YAT-APG------L",
  eic_out = "10YDE-RWENET---I",
  process_type = "A63",
  period_start = lubridate::ymd(x = "2022-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2022-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
