% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hvdc_mode_types}
\alias{hvdc_mode_types}
\title{List HVDC mode types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
hvdc_mode_types
}
\description{
List HVDC mode types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
dplyr::glimpse(entsoeapi::hvdc_mode_types)

}
\keyword{datasets}
