% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hvdc_mode_types}
\alias{hvdc_mode_types}
\title{List HVDC Mode Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
hvdc_mode_types
}
\description{
List HVDC Mode Types
}
\examples{

library(entsoeapi)

str(hvdc_mode_types)

}
\keyword{datasets}
