% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{implicit_offered_transfer_capacities}
\alias{implicit_offered_transfer_capacities}
\title{Get Implicit Offered Transfer Capacities — Day-Ahead & Intraday (11.1)}
\usage{
implicit_offered_transfer_capacities(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(x = Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  contract_type = "A01",
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{eic_out}{Energy Identification Code of the bidding zone
or control area (TSO)}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{contract_type}{Contract market agreement type;
"A01" = Day ahead
"A07" = Intraday
Defaults to "A01" (Day ahead)}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Convenience wrapper that queries the day-ahead (A01) or intraday (A07)
implicit offered transfer capacities.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df1 <- entsoeapi::implicit_offered_transfer_capacities(
  eic_in = "10Y1001A1001A82H",
  eic_out = "10YDK-1--------W",
  period_start = lubridate::ymd(x = "2023-08-16", tz = "CET"),
  period_end = lubridate::ymd(x = "2023-08-17", tz = "CET"),
  contract_type = "A01",
  tidy_output = TRUE
)

dplyr::glimpse(df1)

df2 <- entsoeapi::implicit_offered_transfer_capacities(
  eic_in = "10YFR-RTE------C",
  eic_out = "10YCH-SWISSGRIDZ",
  period_start = lubridate::ymd(x = "2023-08-16", tz = "CET"),
  period_end = lubridate::ymd(x = "2023-08-17", tz = "CET"),
  contract_type = "A07",
  tidy_output = TRUE
)

dplyr::glimpse(df2)
\dontshow{\}) # examplesIf}
}
\seealso{
Other market endpoints: 
\code{\link{aggregated_bids}()},
\code{\link{allocated_transfer_capacities_3rd_countries}()},
\code{\link{already_allocated_total_capacity}()},
\code{\link{auction_revenue}()},
\code{\link{congestion_income}()},
\code{\link{continuous_offered_transfer_capacities}()},
\code{\link{continuous_offered_transfer_capacity}()},
\code{\link{energy_prices}()},
\code{\link{explicit_offered_transfer_capacities}()},
\code{\link{explicit_offered_transfer_capacity}()},
\code{\link{flow_based_allocations}()},
\code{\link{implicit_offered_transfer_capacity}()},
\code{\link{intraday_prices}()},
\code{\link{net_positions}()},
\code{\link{total_nominated_capacity}()}
}
\concept{market endpoints}
