% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicator_types}
\alias{indicator_types}
\title{List Indicator Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2 rows and 3 columns.
}
\usage{
indicator_types
}
\description{
List Indicator Types
}
\examples{

str(entsoeapi::indicator_types)

}
\keyword{datasets}
