% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{intraday_cross_border_transfer_limits}
\alias{intraday_cross_border_transfer_limits}
\title{Get Intraday Cross-Border Transfer Limits (11.3)}
\usage{
intraday_cross_border_transfer_limits(
  eic_in = NULL,
  eic_out = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of in domain}

\item{eic_out}{Energy Identification Code of out domain}

\item{period_start}{the starting date of the in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{the ending date of the outage in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
intraday cross-border transfer limits of DC links
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::intraday_cross_border_transfer_limits(
  eic_in = "10YFR-RTE------C",
  eic_out = "11Y0-0000-0265-K",
  period_start = lubridate::ymd(x = "2023-08-16", tz = "CET"),
  period_end = lubridate::ymd(x = "2023-08-17", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other transmission endpoints: 
\code{\link{costs_of_congestion_management}()},
\code{\link{countertrading}()},
\code{\link{cross_border_physical_flows}()},
\code{\link{day_ahead_commercial_sched}()},
\code{\link{expansion_and_dismantling_project}()},
\code{\link{forecasted_transfer_capacities}()},
\code{\link{net_transfer_capacities}()},
\code{\link{redispatching_cross_border}()},
\code{\link{redispatching_internal}()},
\code{\link{total_commercial_sched}()}
}
\concept{transmission endpoints}
