% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{intraday_cross_border_transfer_limits}
\alias{intraday_cross_border_transfer_limits}
\title{Get Intraday Cross-Border Transfer Limits (11.3)}
\usage{
intraday_cross_border_transfer_limits(
  eic_in = NULL,
  eic_out = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of in domain}

\item{eic_out}{Energy Identification Code of out domain}

\item{period_start}{the starting date of the in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{the ending date of the outage in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A data frame with intraday cross-border transfer limits
}
\description{
intraday cross-border transfer limits of DC links
}
\examples{
df <- entsoeapi::intraday_cross_border_transfer_limits(
  eic_in = "10YFR-RTE------C",
  eic_out = "11Y0-0000-0265-K",
  period_start = lubridate::ymd(x = "2023-08-16", tz = "CET"),
  period_end = lubridate::ymd(x = "2023-08-17", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
