% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_load.R
\name{load_year_ahead_forecast_margin}
\alias{load_year_ahead_forecast_margin}
\title{Get Year-Ahead Forecast Margin (8.1)}
\usage{
load_year_ahead_forecast_margin(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the bidding zone/
country/control area}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
It is defined as a difference between yearly forecast of
available generation capacity and yearly forecast of
total load, taking into account the forecast of
total generation capacity forecast of availability of
generation and forecast of reserves contracted for
system services.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::load_year_ahead_forecast_margin(
  eic = "10Y1001A1001A82H",
  period_start = lubridate::ymd(x = "2019-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2019-12-31", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other load endpoints: 
\code{\link{load_actual_total}()},
\code{\link{load_day_ahead_total_forecast}()},
\code{\link{load_month_ahead_total_forecast}()},
\code{\link{load_week_ahead_total_forecast}()},
\code{\link{load_year_ahead_total_forecast}()}
}
\concept{load endpoints}
