% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{location_eic}
\alias{location_eic}
\title{Get Location_V EIC codes}
\usage{
location_eic()
}
\description{
This function downloads approved Location_V Energy Identification Codes from
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes/
}
\examples{
eic_location <- location_eic()
}
