% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{market_product_types}
\alias{market_product_types}
\title{List Market Product Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 14 rows and 3 columns.
}
\usage{
market_product_types
}
\description{
List Market Product Types
}
\examples{

str(entsoeapi::market_product_types)

}
\keyword{datasets}
