% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{message_types}
\alias{message_types}
\title{List Message Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 150 rows and 3 columns.
}
\usage{
message_types
}
\description{
List Message Types
}
\examples{

library(entsoeapi)

str(message_types)

}
\keyword{datasets}
