% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_market.R
\name{net_positions}
\alias{net_positions}
\title{Get Implicit Auction — Net Positions (12.1.E)}
\usage{
net_positions(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  contract_type = "A01",
  tidy_output = FALSE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the bidding zone}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{contract_type}{Contract market agreement type;
"A01" = Day ahead
"A07" = Intraday
Defaults to "A01" (Day ahead)}

\item{tidy_output}{Defaults to FALSE.
If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Net positions resulting from implicit auctions per bidding zone.
}
\examples{
df <- entsoeapi::net_positions(
  eic = "10YCZ-CEPS-----N",
  period_start = lubridate::ymd(x = "2015-12-31", tz = "CET"),
  period_end = lubridate::ymd(x = "2016-01-01", tz = "CET"),
  contract_type = "A01",
  tidy_output = TRUE
)

str(df)

}
