% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_transmission.R
\name{net_transfer_capacities}
\alias{net_transfer_capacities}
\title{Get Net Transfer Capacities — Week/Month/Year Ahead (11.1)}
\usage{
net_transfer_capacities(
  eic_in = NULL,
  eic_out = NULL,
  contract_type = "A02",
  period_start = ymd(Sys.Date() - days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic_in}{Energy Identification Code of in domain}

\item{eic_out}{Energy Identification Code of out domain}

\item{contract_type}{Contract market agreement type:
"A02" = Week ahead
"A03" = Month ahead
"A04" = Year ahead
Defaults to "A02" (Week ahead)}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Estimated net transfer capacities for the week-ahead,
month-ahead, or year-ahead time frames. For day-ahead
NTC, use \code{\link[=forecasted_transfer_capacities]{forecasted_transfer_capacities()}} instead.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::net_transfer_capacities(
  eic_in = "10YCZ-CEPS-----N",
  eic_out = "10YSK-SEPS-----K",
  contract_type = "A02",
  period_start = lubridate::ymd(x = "2024-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-02-01", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other transmission endpoints: 
\code{\link{costs_of_congestion_management}()},
\code{\link{countertrading}()},
\code{\link{cross_border_physical_flows}()},
\code{\link{day_ahead_commercial_sched}()},
\code{\link{expansion_and_dismantling_project}()},
\code{\link{forecasted_transfer_capacities}()},
\code{\link{intraday_cross_border_transfer_limits}()},
\code{\link{redispatching_cross_border}()},
\code{\link{redispatching_internal}()},
\code{\link{total_commercial_sched}()}
}
\concept{transmission endpoints}
