% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{netted_volumes}
\alias{netted_volumes}
\title{Get Netted Volumes (IFs IN 3.10)}
\usage{
netted_volumes(
  eic = NULL,
  period_start = ymd(Sys.Date() - days(x = 7L), tz = "CET"),
  period_end = ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the area}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One day range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One day range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
The net position informs whether the given area imports or exports energy.
Those rows which hold the queried eic value in the
'ts_connecting_domain_mrid' column show the export value.
Those rows which hold the queried eic value in the
'ts_acquiring_domain_mrid' column show the import value.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::netted_volumes(
  eic = "10YCZ-CEPS-----N",
  period_start = lubridate::ymd(x = "2022-08-16", tz = "CET"),
  period_end = lubridate::ymd(x = "2022-08-17", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other balancing endpoints: 
\code{\link{activated_balancing_prices}()},
\code{\link{aggregated_balancing_energy_bids}()},
\code{\link{allocation_of_cross_zonal_balancing_cap}()},
\code{\link{balancing_border_cap_limit}()},
\code{\link{balancing_energy_bids}()},
\code{\link{changes_to_bid_availability}()},
\code{\link{contracted_reserves}()},
\code{\link{current_balancing_state}()},
\code{\link{elastic_demands}()},
\code{\link{exchanged_volumes}()},
\code{\link{exchanged_volumes_per_border}()},
\code{\link{fcr_total_capacity}()},
\code{\link{financial_expenses_and_income}()},
\code{\link{hvdc_link_constrains}()},
\code{\link{imbalance_prices}()},
\code{\link{imbalance_volumes}()},
\code{\link{netted_volumes_per_border}()},
\code{\link{procured_balancing_capacity}()},
\code{\link{rr_and_frr_actual_capacity}()},
\code{\link{shares_of_fcr_capacity}()},
\code{\link{sharing_of_rr_and_frr_capacity}()}
}
\concept{balancing endpoints}
