% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{netted_volumes_per_border}
\alias{netted_volumes_per_border}
\title{Get Netted & Exchanged Volumes Per Border (IFs 3.10, 3.16 & 3.17)}
\usage{
netted_volumes_per_border(
  acquiring_eic = NULL,
  connecting_eic = NULL,
  process_type = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{acquiring_eic}{Energy Identification Code of the acquiring area}

\item{connecting_eic}{Energy Identification Code of the connecting area}

\item{process_type}{type of balancing process
"A51" aFRR
"A60" mFRR with scheduled activation
"A61" mFRR with direct activation
"A63" Imbalance Netting}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One day range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One day range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Netted and exchanged balancing energy volumes on a specific border.
Covers imbalance netting (3.10), aFRR exchange (3.16), and
mFRR exchange (3.17) per border between an acquiring and a
connecting domain.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
\dontrun{
df <- entsoeapi::netted_volumes_per_border(
  acquiring_eic = "10YBE----------2",
  connecting_eic = "10YFR-RTE------C",
  process_type = "A63",
  period_start = lubridate::ymd(x = "2025-03-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2025-03-02", tz = "CET"),
  tidy_output = TRUE
)

dplyr::glimpse(df)
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other balancing endpoints: 
\code{\link{activated_balancing_prices}()},
\code{\link{aggregated_balancing_energy_bids}()},
\code{\link{allocation_of_cross_zonal_balancing_cap}()},
\code{\link{balancing_border_cap_limit}()},
\code{\link{balancing_energy_bids}()},
\code{\link{changes_to_bid_availability}()},
\code{\link{contracted_reserves}()},
\code{\link{current_balancing_state}()},
\code{\link{elastic_demands}()},
\code{\link{exchanged_volumes}()},
\code{\link{exchanged_volumes_per_border}()},
\code{\link{fcr_total_capacity}()},
\code{\link{financial_expenses_and_income}()},
\code{\link{hvdc_link_constrains}()},
\code{\link{imbalance_prices}()},
\code{\link{imbalance_volumes}()},
\code{\link{netted_volumes}()},
\code{\link{procured_balancing_capacity}()},
\code{\link{rr_and_frr_actual_capacity}()},
\code{\link{shares_of_fcr_capacity}()},
\code{\link{sharing_of_rr_and_frr_capacity}()}
}
\concept{balancing endpoints}
