% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{object_aggregation_types}
\alias{object_aggregation_types}
\title{List object aggregation types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 16 rows and 3 columns.
}
\usage{
object_aggregation_types
}
\description{
List object aggregation types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
dplyr::glimpse(entsoeapi::object_aggregation_types)

}
\keyword{datasets}
