% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{object_aggregation_types}
\alias{object_aggregation_types}
\title{List Object Aggregation Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 16 rows and 3 columns.
}
\usage{
object_aggregation_types
}
\description{
List Object Aggregation Types
}
\examples{

library(entsoeapi)

str(object_aggregation_types)

}
\keyword{datasets}
