% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages.R
\name{outages_both}
\alias{outages_both}
\title{Get Unavailability of Production & Generation Units (15.1.A&B + 15.1.C&D)}
\usage{
outages_both(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() + lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date() + lubridate::days(x = 2L), tz = "CET"),
  doc_status = NULL,
  event_nature = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the bidding zone/control area
(To extract outages of bidding zone DE-AT-LU area,
it is recommended to send queries per control area
i.e. CTA|DE(50Hertz), CTA|DE(Amprion), CTA|DE(TeneTGer),
CTA|DE(TransnetBW),CTA|AT,CTA|LU but not per bidding zone.)}

\item{period_start}{the starting date of the in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{the ending date of the outage in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{doc_status}{Notification document status. "A05" for active,
"A09" for cancelled and "A13" for withdrawn.
Defaults to NULL which means "A05" and "A09" together.}

\item{event_nature}{"A53" for planned maintenance.
"A54" for unplanned outage.
Defaults to NULL which means both of them.}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
The planned and forced unavailability of production and
generation units expected to last at least one market time unit
up to 3 years ahead. The "available capacity during the event"
means the minimum available generation capacity during the period specified.
}
\examples{
df <- entsoeapi::outages_both(
  eic = "10YFR-RTE------C",
  period_start = lubridate::ymd(
                   x = Sys.Date() + lubridate::days(x = 1L),
                   tz = "CET"
                 ),
  period_end = lubridate::ymd(
                 x = Sys.Date() + lubridate::days(x = 2L),
                 tz = "CET"
               )
)

str(df)

}
