% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages.R
\name{outages_cons_units}
\alias{outages_cons_units}
\title{Get Unavailability of Consumption Units. (7.1.A&B)}
\usage{
outages_cons_units(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() + lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date() + lubridate::days(x = 2L), tz = "CET"),
  period_start_update = NULL,
  period_end_update = NULL,
  doc_status = NULL,
  event_nature = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the bidding zone/
control area
(To extract outages of bidding zone DE-AT-LU area,
it is recommended to send queries per control area
i.e. CTA|DE(50Hertz), CTA|DE(Amprion), CTA|DE(TeneTGer),
CTA|DE(TransnetBW),CTA|AT,CTA|LU but not per bidding zone.)}

\item{period_start}{the starting date of the in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{the ending date of the outage in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_start_update}{notification submission/update starting date
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end_update}{notification submission/update ending date
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{doc_status}{Notification document status. "A05" for active,
"A09" for cancelled and "A13" for withdrawn.
Defaults to NULL which means "A05" and "A09" together.}

\item{event_nature}{"A53" for planned maintenance.
"A54" for unplanned outage.
Defaults to NULL which means both of them.}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Unavailability of consumption units in aggregated form.
All planned and forced outages in selected area are
aggregated according to the market time unit. The list of
specific consumption units are not provided.
}
\examples{
df <- entsoeapi::outages_cons_units(
  eic          = "10YFI-1--------U",
  period_start = lubridate::ymd(x = "2024-04-10", tz = "CET"),
  period_end   = lubridate::ymd(x = "2024-04-11", tz = "CET")
)

str(df)

}
