% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages.R
\name{outages_fallbacks}
\alias{outages_fallbacks}
\title{Get Fall-Back Procedures. (IFs IN 7.2, mFRR 3.11, aFRR 3.10)}
\usage{
outages_fallbacks(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  process_type = "A63",
  event_nature = "A53",
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the bidding zone/
control area}

\item{period_start}{the starting date of the in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{the ending date of the outage in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{process_type}{"A47" = mFRR
"A51" = aFRR
"A63" = imbalance netting
defaults to "A63"}

\item{event_nature}{"C47" = Disconnection,
"A53" = Planned maintenance,
"A54": Unplanned outage,
"A83" = Auction cancellation (used in case
no solution found or algorithm failure);
Defaults to "A53".}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
It publishes of application of fall back procedures by participants
in European platforms as a result of disconnection of DSO from
the European platform, unavailability of European platform itself
(planned or unplanned outage) or the situation where the algorithm
used on the platform fails or does not find solution.
}
\examples{

library(entsoeapi)
library(lubridate)

df <- outages_fallbacks(eic          = "10YBE----------2",
                        period_start = ymd(x = "2023-01-01", tz = "CET"),
                        period_end   = ymd(x = "2024-01-01", tz = "CET"),
                        process_type = "A51",
                        event_nature   = "C47")

str(df)

}
