% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages.R
\name{outages_offshore_grid}
\alias{outages_offshore_grid}
\title{Get Unavailability of Offshore Grid Infrastructure. (10.1.C)}
\usage{
outages_offshore_grid(
  eic = NULL,
  period_start = ymd(Sys.Date() + days(x = 1L), tz = "CET"),
  period_end = ymd(Sys.Date() + days(x = 2L), tz = "CET"),
  doc_status = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the bidding zone/
control area (To extract outages of bidding zone
DE-AT-LU area, it is recommended to send queries
per control area i.e. CTA|DE(50Hertz), CTA|DE(Amprion),
CTA|DE(TeneTGer), CTA|DE(TransnetBW),CTA|AT,CTA|LU
but not per bidding zone.)}

\item{period_start}{the starting date of the in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{the ending date of the outage in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{doc_status}{Notification document status. NULL or "A05"
for active and "A13" for withdrawn.
Defaults to NULL.}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with the queried data.
}
\description{
Unavailability of the off-shore grid that reduce wind power
feed-in during at least one market time unit.
Wind power fed in at the time of the change in the availability
is provided.
}
\examples{
\dontshow{if (there_is_provider() && nchar(Sys.getenv("ENTSOE_PAT")) > 0L) withAutoprint(\{ # examplesIf}
df <- entsoeapi::outages_offshore_grid(
  eic = "10Y1001A1001A82H",
  period_start = lubridate::ymd(
    x = Sys.Date() - lubridate::days(x = 30L),
    tz = "CET"
  ),
  period_end = lubridate::ymd(x = Sys.Date(), tz = "CET")
)

dplyr::glimpse(df)
\dontshow{\}) # examplesIf}
}
\seealso{
Other outage endpoints: 
\code{\link{outages_both}()},
\code{\link{outages_cons_units}()},
\code{\link{outages_fallbacks}()},
\code{\link{outages_gen_units}()},
\code{\link{outages_prod_units}()},
\code{\link{outages_transmission_grid}()}
}
\concept{outage endpoints}
