% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_outages.R
\name{outages_offshore_grid}
\alias{outages_offshore_grid}
\title{Get Unavailability of Offshore Grid Infrastructure. (10.1.A&B)}
\usage{
outages_offshore_grid(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() + lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date() + lubridate::days(x = 2L), tz = "CET"),
  period_start_update = NULL,
  period_end_update = NULL,
  doc_status = NULL,
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the bidding zone/
control area (To extract outages of bidding zone
DE-AT-LU area, it is recommended to send queries
per control area i.e. CTA|DE(50Hertz), CTA|DE(Amprion),
CTA|DE(TeneTGer), CTA|DE(TransnetBW),CTA|AT,CTA|LU
but not per bidding zone.)}

\item{period_start}{the starting date of the in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{the ending date of the outage in-scope period
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_start_update}{notification submission/update starting date
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end_update}{notification submission/update ending date
in POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{doc_status}{Notification document status. "A05" for active,
"A09" for cancelled and "A13" for withdrawn.
Defaults to NULL which means "A05" and "A09" together.}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Unavailability of the off-shore grid that reduce wind power
feed-in during at least one market time unit.
Wind power fed in at the time of the change in the availability
is provided.
}
\examples{

library(entsoeapi)
library(lubridate)

df <- outages_offshore_grid(
  eic                 = "10Y1001A1001A82H",
  period_start        = ymd(x = Sys.Date() - days(x = 365L), tz = "CET"),
  period_end          = ymd(x = Sys.Date(), tz = "CET"),
  period_start_update = ymd(x = Sys.Date() - days(x = 365L), tz = "CET"),
  period_end_update   = ymd(x = Sys.Date(), tz = "CET")
)

str(df)

}
