% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{party_eic}
\alias{party_eic}
\title{Get Party_X EIC codes}
\usage{
party_eic()
}
\description{
This function downloads approved Party_X Energy Identification Codes from
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes/
}
\examples{
eic_party <- party_eic()
}
