% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{party_eic}
\alias{party_eic}
\title{Get Party_X Energy Identification Codes}
\usage{
party_eic()
}
\value{
A tibble of accordingly filtered EIC codes, which contains such columns as
\code{EicCode}, \code{EicDisplayName}, \code{EicLongName}, \code{EicParent},
\code{EicResponsibleParty}, \code{EicStatus}, \code{MarketParticipantPostalCode},
\code{MarketParticipantIsoCountryCode}, \code{MarketParticipantVatCode},
\code{EicTypeFunctionList} and \code{type}.
}
\description{
This function downloads approved party X
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
It covers market participants.
}
\examples{

library(entsoeapi)

eic_party <- party_eic()

str(eic_party)

}
