% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{payment_terms_types}
\alias{payment_terms_types}
\title{List Payment Terms Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
payment_terms_types
}
\description{
List Payment Terms Types
}
\examples{

library(entsoeapi)

str(payment_terms_types)

}
\keyword{datasets}
