% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_category_types}
\alias{price_category_types}
\title{List Price Category Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 7 rows and 3 columns.
}
\usage{
price_category_types
}
\description{
List Price Category Types
}
\examples{

library(entsoeapi)

str(price_category_types)

}
\keyword{datasets}
