% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_component_types}
\alias{price_component_types}
\title{List Price Component Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
price_component_types
}
\description{
List Price Component Types
}
\examples{

library(entsoeapi)

str(price_component_types)

}
\keyword{datasets}
