% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{price_direction_types}
\alias{price_direction_types}
\title{List Price Direction Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2 rows and 3 columns.
}
\usage{
price_direction_types
}
\description{
List Price Direction Types
}
\examples{

str(entsoeapi::price_direction_types)

}
\keyword{datasets}
