% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{process_types}
\alias{process_types}
\title{List Process Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 51 rows and 3 columns.
}
\usage{
process_types
}
\description{
List Process Types
}
\examples{

library(entsoeapi)

str(process_types)

}
\keyword{datasets}
