% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{procured_balancing_capacity}
\alias{procured_balancing_capacity}
\title{Get Procured Balancing Capacity (GL EB 12.3.F)}
\usage{
procured_balancing_capacity(
  eic = NULL,
  process_type = NULL,
  type_market_agreement = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the area domain}

\item{process_type}{type of frequency restoration reserve
"A51" aFRR
"A52" FCR
"A47" mFRR}

\item{type_market_agreement}{Optional market agreement type code}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Procured balancing capacity by TSOs.
100 documents limit applies; paging is handled transparently.
}
\examples{
df <- entsoeapi::procured_balancing_capacity(
  eic = "10YCZ-CEPS-----N",
  process_type = "A51",
  period_start = lubridate::ymd(x = "2024-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2024-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
