% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quality_types}
\alias{quality_types}
\title{List Quality Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 5 rows and 3 columns.
}
\usage{
quality_types
}
\description{
List Quality Types
}
\examples{

library(entsoeapi)

str(quality_types)

}
\keyword{datasets}
