% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reason_code_types}
\alias{reason_code_types}
\title{List reason code types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 166 rows and 3 columns.
}
\usage{
reason_code_types
}
\description{
List reason code types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
dplyr::glimpse(entsoeapi::reason_code_types)

}
\keyword{datasets}
