% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reason_code_types}
\alias{reason_code_types}
\title{List Reason Code Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 106 rows and 3 columns.
}
\usage{
reason_code_types
}
\description{
List Reason Code Types
}
\examples{

library(entsoeapi)

str(reason_code_types)

}
\keyword{datasets}
