% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_congestion.R
\name{redispatching_internal}
\alias{redispatching_internal}
\title{Get Redispatching Internal (13.1.A)}
\usage{
redispatching_internal(
  eic = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 1L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the control area}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE. If TRUE, then flatten nested tables.}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Changes in production and load (increase or decrease) to
relieve internal congestion lines that exceeds its capacity.
100 documents limit applies!!
}
\examples{

library(entsoeapi)
library(lubridate)

# Netherlands' internal redispatching.
df <- redispatching_internal(
  eic          = "10YNL----------L",
  period_start = ymd(x = "2023-11-01", tz = "CET"),
  period_end   = ymd(x = "2023-12-01", tz = "CET"),
  tidy_output  = TRUE
)

str(df)

}
