% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{resource_object_eic}
\alias{resource_object_eic}
\title{Get Resource_Object_W EIC codes}
\usage{
resource_object_eic()
}
\description{
This function downloads approved Resource_Object_W Energy Identification Codes from
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes/
}
\examples{
eic_resource_object <- resource_object_eic()
}
