% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{resource_object_eic}
\alias{resource_object_eic}
\title{Get Resource Object W Energy Identification Codes}
\usage{
resource_object_eic()
}
\value{
A tibble of accordingly filtered EIC codes, which contains such columns as
\code{EicCode}, \code{EicDisplayName}, \code{EicLongName}, \code{EicParent},
\code{EicResponsibleParty}, \code{EicStatus}, \code{MarketParticipantPostalCode},
\code{MarketParticipantIsoCountryCode}, \code{MarketParticipantVatCode},
\code{EicTypeFunctionList} and \code{type}.
}
\description{
This function downloads approved resource object W
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
A resource that can either produce or consume energy
and that is reported in a schedule.
}
\examples{
eic_resource_object <- entsoeapi::resource_object_eic()

str(eic_resource_object)

}
