% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rights_types}
\alias{rights_types}
\title{List Rights Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 6 rows and 3 columns.
}
\usage{
rights_types
}
\description{
List Rights Types
}
\examples{

library(entsoeapi)

str(rights_types)

}
\keyword{datasets}
