% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{role_types}
\alias{role_types}
\title{List Role Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 59 rows and 3 columns.
}
\usage{
role_types
}
\description{
List Role Types
}
\examples{

library(entsoeapi)

str(role_types)

}
\keyword{datasets}
