% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{rr_and_frr_actual_capacity}
\alias{rr_and_frr_actual_capacity}
\title{Get RR and FRR Actual Capacity (SO GL 188.4 & 189.3)}
\usage{
rr_and_frr_actual_capacity(
  eic = NULL,
  process_type = NULL,
  business_type = "C78",
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the area domain}

\item{process_type}{type of reserve
"A56" FRR
"A46" RR}

\item{business_type}{type of aggregation
"C77" minimum
"C78" average
"C79" max
defaults to "C78"}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Actual capacity of Frequency Restoration Reserve (FRR) or
Replacement Reserve (RR).
}
\examples{
df <- entsoeapi::rr_and_frr_actual_capacity(
  eic = "10YAT-APG------L",
  process_type = "A56",
  period_start = lubridate::ymd(x = "2022-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2022-04-01", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
