% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{settlement_method_types}
\alias{settlement_method_types}
\title{List Settlement Method Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2 rows and 3 columns.
}
\usage{
settlement_method_types
}
\description{
List Settlement Method Types
}
\examples{

str(entsoeapi::settlement_method_types)

}
\keyword{datasets}
