% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_balancing.R
\name{shares_of_fcr_capacity}
\alias{shares_of_fcr_capacity}
\title{Get Shares of FCR Capacity (SO GL 187.2)}
\usage{
shares_of_fcr_capacity(
  eic = NULL,
  business_type = NULL,
  period_start = lubridate::ymd(Sys.Date() - lubridate::days(x = 7L), tz = "CET"),
  period_end = lubridate::ymd(Sys.Date(), tz = "CET"),
  tidy_output = TRUE,
  security_token = Sys.getenv("ENTSOE_PAT")
)
}
\arguments{
\item{eic}{Energy Identification Code of the area domain}

\item{business_type}{type of FCR capacity share
"C23" share of FCR capacity
"B95" contracted FCR capacity}

\item{period_start}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{period_end}{POSIXct or YYYY-MM-DD HH:MM:SS format
One year range limit applies}

\item{tidy_output}{Defaults to TRUE. flatten nested tables}

\item{security_token}{Security token for ENTSO-E transparency platform}
}
\description{
Shares of Frequency Containment Reserve capacity per area.
One year range limit applies.
}
\examples{
df <- entsoeapi::shares_of_fcr_capacity(
  eic = "10YDE-VE-------2",
  business_type = "C23",
  period_start = lubridate::ymd(x = "2022-01-01", tz = "CET"),
  period_end = lubridate::ymd(x = "2022-01-02", tz = "CET"),
  tidy_output = TRUE
)

str(df)

}
