% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status_types}
\alias{status_types}
\title{List status types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 76 rows and 3 columns.
}
\usage{
status_types
}
\description{
List status types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
dplyr::glimpse(entsoeapi::status_types)

}
\keyword{datasets}
