% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status_types}
\alias{status_types}
\title{List Status Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 16 rows and 3 columns.
}
\usage{
status_types
}
\description{
List Status Types
}
\examples{

library(entsoeapi)

str(status_types)

}
\keyword{datasets}
