% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{substation_eic}
\alias{substation_eic}
\title{Get Substation A Energy Identification Codes}
\usage{
substation_eic()
}
\value{
A tibble of accordingly filtered EIC codes, which contains such columns as
\code{EicCode}, \code{EicDisplayName}, \code{EicLongName}, \code{EicParent},
\code{EicResponsibleParty}, \code{EicStatus}, \code{MarketParticipantPostalCode},
\code{MarketParticipantIsoCountryCode}, \code{MarketParticipantVatCode},
\code{EicTypeFunctionList} and \code{type}.
}
\description{
This function downloads all approved substation A
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
Substation is a facility equipment that steps up or
steps down the voltage in utility power lines.
Voltage is stepped up where power is sent through
long distance transmission lines, and stepped down
where the power is to enter the local distribution lines.
They can be classified as normal outside substation,
armoured substation and underground substation.
}
\examples{
eic_substation <- entsoeapi::substation_eic()

str(eic_substation)

}
