% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tariff_types}
\alias{tariff_types}
\title{List Tariff Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 7 rows and 3 columns.
}
\usage{
tariff_types
}
\description{
List Tariff Types
}
\examples{

library(entsoeapi)

str(tariff_types)

}
\keyword{datasets}
