% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{tie_line_eic}
\alias{tie_line_eic}
\title{Get Tie_Line_T EIC codes}
\usage{
tie_line_eic()
}
\description{
This function downloads approved Tie_Line_T Energy Identification Codes from
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes/
}
\examples{
eic_tie_line <- tie_line_eic()
}
