% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_helpers.R
\name{tie_line_eic}
\alias{tie_line_eic}
\title{Get Tie Line T Energy Identification Codes}
\usage{
tie_line_eic()
}
\description{
This function downloads approved tie line T
energy identification codes from this site:
https://www.entsoe.eu/data/energy-identification-codes-eic/eic-approved-codes
It covers a transmission line that connects different areas
excluding HVDC interconnectors.
}
\examples{

library(entsoeapi)

eic_tie_line <- tie_line_eic()

str(eic_tie_line)

}
