% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeframe_types}
\alias{timeframe_types}
\title{List timeframe types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 46 rows and 3 columns.
}
\usage{
timeframe_types
}
\description{
List timeframe types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
dplyr::glimpse(entsoeapi::timeframe_types)

}
\keyword{datasets}
