% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transmission_pair_eic_dict}
\alias{transmission_pair_eic_dict}
\title{List Transmission Pair EIC Dictionary}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 114 rows and 8 columns.
}
\usage{
transmission_pair_eic_dict
}
\description{
List Transmission Pair EIC Dictionary
}
\examples{

library(entsoeapi)

str(transmission_pair_eic_dict)

}
\keyword{datasets}
