% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_multiplier}
\alias{unit_multiplier}
\title{List Unit Multiplier}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
unit_multiplier
}
\description{
List Unit Multiplier
}
\examples{

library(entsoeapi)

str(unit_multiplier)

}
\keyword{datasets}
