% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_of_measure_types}
\alias{unit_of_measure_types}
\title{List unit of measure types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 33 rows and 3 columns.
}
\usage{
unit_of_measure_types
}
\description{
List unit of measure types based on
'https://www.entsoe.eu/Documents/EDI/Library/CodelistV94.zip'
}
\examples{
dplyr::glimpse(entsoeapi::unit_of_measure_types)

}
\keyword{datasets}
