% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_of_measure_types}
\alias{unit_of_measure_types}
\title{List Unit of Measure Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 14 rows and 3 columns.
}
\usage{
unit_of_measure_types
}
\description{
List Unit of Measure Types
}
\examples{

library(entsoeapi)

str(unit_of_measure_types)

}
\keyword{datasets}
