% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_symbol_types}
\alias{unit_symbol_types}
\title{List Unit Symbol Types}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11 rows and 3 columns.
}
\usage{
unit_symbol_types
}
\description{
List Unit Symbol Types
}
\examples{

library(entsoeapi)

str(unit_symbol_types)

}
\keyword{datasets}
